==> 3a_descriptive_analysis_all_raw_data.Rmd <==
---
title: 'Descriptive analysis: Air pollution, noise level and allostatic load'
author: "Yuchan Mou"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r set_up, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
options(knitr.table.format = "html")
```

```{r load packages, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
# load pkgs
packages <- c(
  "dplyr", "tidyverse", "readxl", "openxlsx", "ggplot2", "kableExtra",
  "ggcorrplot", "JointAI", "patchwork"
)
sapply(packages, library, character.only = T)
```

```{r import data3, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
load("/Users/ymou/helix_project/data/analysis_data/df_al_final.RData")
Sys.Date()
```

```{r define variables, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
expo <- c(
  "hs_no2_yr_hs_t", "hs_no2_yr_hs_h", "hs_no2_yr_hs_s", "hs_no2_yr_hs_r", "hs_no2_yr_hs_p",
  "hs_pm25_yr_hs_t", "hs_pm25_yr_hs_h", "hs_pm25_yr_hs_s", "hs_pm25_yr_hs_r", "hs_pm25_yr_hs_p",
  "hs_pm25abs_yr_hs_t", "hs_pm25abs_yr_hs_h", "hs_pm25abs_yr_hs_s", "hs_pm25abs_yr_hs_r", "hs_pm25abs_yr_hs_p",
  "hs_pm10_yr_hs_t", "hs_pm10_yr_hs_h", "hs_pm10_yr_hs_s", "hs_pm10_yr_hs_r", "hs_pm10_yr_hs_p",
  "hs_lden_c_h", "hs_lden_c_s", "hs_lden_t"
)

cov <- c(
  "e3_sex", "hs_mvpa", "hs_sd_wk", "hs_child_age_years", # child characteristics
  "FAS_cat", "h_native", "h_ethnicity_cauc", "h_edumc", "h_edufc",
  "hs_smk_parents", "hs_globalexp", "e3_asmokyn_p", "e3_alcpreg_yn", "h_marital", "h_age", "h_parity"
)

out <- c("AL_2_tot", "AL_2_tot_wavg", "AL_z2_tot", "AL_z2_tot_wavg") # sex-specific scores
# "AL_1_tot", "AL_1_tot_wavg", "AL_z1_tot", "AL_z1_tot_wavg": score based on whole sample

cardio <- c("hs_zsys_bp", "hs_zdia_bp", "bp_pulse_sd")
metab <- c(
  "hs_zbmi_who", "hs_zwaist_mets", "hs_skf_sum2", "hs_fatprop_bia",
  "hs_hdlchol_c", "hs_non_hdl_c", "hs_triglyc_c", "Leptin", "Adiponectin"
) # "hs_ldlchol_c", "hs_totchol_c"
immune <- c("CRP", "IL1beta", "IL6", "IL8", "IL10", "TNFalfa")
neuro <- c("cortisol_prod_log10")
```

# Descriptive analysis of exposure and covariates

## 1. Distribution

The plots below shows the distributions of exposures and covariates.

### - NO2

```{r no2, fig.width = 10 , fig.height = 10}
df_plot <- df_final %>%
  select(starts_with(c("hs_no2", "h_no2")), "cohort") %>%
  as.data.frame()

par(mfrow = c(4, 4))

plot_all(df_plot)

# separate by cohort
df_plot %>%
  pivot_longer(cols = 1:ncol(df_plot) - 1) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x") +
  ggtitle("By cohort")
```

### - PM25

```{r pm25, fig.width = 10 , fig.height = 10}
df_plot <- df_final %>%
  select(starts_with(c("hs_pm25_", "h_pm25_")), "cohort") %>%
  as.data.frame()

par(mfrow = c(4, 4))

plot_all(df_plot)

# separate by cohort
df_plot %>%
  pivot_longer(cols = 1:ncol(df_plot) - 1) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x") +
  ggtitle("By cohort")
```

### - PM25abs

```{r pm25abs, fig.width = 10 , fig.height = 10}
df_plot <- df_final %>%
  select(starts_with(c("hs_pm25abs", "h_pm25abs")), "cohort") %>%
  as.data.frame()

par(mfrow = c(4, 4))

plot_all(df_plot)

# separate by cohort
df_plot %>%
  pivot_longer(cols = 1:ncol(df_plot) - 1) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x") +
  ggtitle("By cohort")
```

### - PM10

```{r pm10, fig.width = 10 , fig.height = 10}
df_plot <- df_final %>%
  select(starts_with(c("hs_pm10", "h_pm10")), "cohort") %>%
  as.data.frame()

par(mfrow = c(4, 4))

plot_all(df_plot)

# separate by cohort
df_plot %>%
  pivot_longer(cols = 1:ncol(df_plot) - 1) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x") +
  ggtitle("By cohort")
```

### - Lden

```{r lden, fig.width = 13 , fig.height = 13}
df_plot <- df_final %>%
  select(starts_with(c("hs_lden")), "cohort") %>%
  as.data.frame()

par(mfrow = c(4, 4))

plot_all(df_plot)

# separate by cohort
df_plot_continuous <- df_plot %>%
  select(-c("hs_lden_c_h", "hs_lden_c_s"))
df_plot_continuous %>%
  pivot_longer(cols = 1:ncol(df_plot_continuous) - 1) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x") +
  ggtitle("By cohort")

p1 <- df_final %>%
  ggplot(aes(x = hs_lden_c_h, fill = cohort)) +
  geom_bar(position = "dodge") +
  ggtitle("Lden Home by cohort")

p2 <- df_final %>%
  ggplot(aes(x = hs_lden_c_s, fill = cohort)) +
  geom_bar(position = "dodge") +
  ggtitle("Lden Home by cohort")

wrap_plots(p1, p2, ncol = 2, nrow = 1)
```

### - Covariates

```{r cov, fig.width = 20 , fig.height = 20}
df_plot <- df_final %>%
  select(
    "hs_child_age_years", "child_age_grp",
    "FAS_score", "FAS_cat",
    "h_native", "h_ethnicity_c", "h_ethnicity_cauc", "h_edumc", "h_edufc",
    "e3_sex", "hs_mvpa", "hs_sd_wk",
    "hs_smk_parents", "hs_globalexp", "e3_asmokyn_p", "e3_alcpreg_yn",
    "h_marital", "h_parity", "cohort"
  ) %>%
  # mutate_at(vars(-cohort), as.numeric) %>%
  as.data.frame()

par(mfrow = c(5, 3))

plot_all(df_plot)
```

```{r cov_stratification, fig.width = 15 , fig.height = 15}
p1 <- df_final %>%
  ggplot(aes(x = child_age_grp, fill = cohort)) +
  geom_bar(position = "dodge") +
  ggtitle("Age group by cohort")

p2 <- df_final %>%
  ggplot(aes(x = e3_sex, fill = child_age_grp)) +
  geom_bar(position = "dodge") +
  ggtitle("Child sex by age group")

p3 <- df_final %>%
  ggplot(aes(x = h_ethnicity_c, fill = child_age_grp)) +
  geom_bar(position = "dodge") +
  ggtitle("Child ethnicity by age group")

p4 <- df_final %>%
  ggplot(aes(x = h_ethnicity_c, fill = cohort)) +
  geom_bar(position = "dodge") +
  ggtitle("Child ethnicity by cohort")

p5 <- df_final %>%
  ggplot(aes(x = h_ethnicity_c, fill = FAS_cat)) +
  geom_bar(position = "dodge") +
  ggtitle("Child ethnicity by family affulent scale")

wrap_plots(p1, p2, p3, p4, p5, ncol = 2, nrow = 3)
```

## 2. Complete cases of exposures

```{r complete_expo}
expo_desc <- df_final %>%
  filter(!is.na(e3_sex)) %>%
  select(cohort, all_of(expo)) %>%
  group_by(cohort)

as.data.frame(cbind(
  "#" = colSums(!is.na(expo_desc))
)) %>%
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T) %>%
  scroll_box(fixed_thead = T)

expo_desc_cohort <- group_split(expo_desc)
cohort <- group_keys(expo_desc)

exposure_desc <- as.data.frame(cbind(
  "# SAB" = colSums(!is.na(expo_desc_cohort[[1]])),
  "# EDEN" = colSums(!is.na(expo_desc_cohort[[2]])),
  "# BIB" = colSums(!is.na(expo_desc_cohort[[3]])),
  "# RHEA" = colSums(!is.na(expo_desc_cohort[[4]])),
  "# KANC" = colSums(!is.na(expo_desc_cohort[[5]])),
  "# MOBA" = colSums(!is.na(expo_desc_cohort[[6]]))
))

exposure_desc %>%
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T) %>%
  scroll_box(fixed_thead = T)
```

## 3. Missing percentage of all variables

```{r missing,  fig.width = 15 , fig.height = 15}
daf_impcheck <- df_final %>%
  select(starts_with(c(
    "hs_no2", "h_no2", "hs_pm25", "h_pm25", "hs_pm25abs", "h_pm25abs", "hs_pm10", "h_pm10", "hs_lden",
    "FAS_score", "FAS_cat", "h_subjincome", "familyincomeC", "e3_ses",
    "h_native", "h_ethnicity_c", "h_ethnicity_cauc", "h_edumc", "h_edufc",
    "e3_sex", "hs_mvpa",
    "hs_smk_parents", "hs_globalexp", "e3_asmokyn_p", "e3_alcpreg_yn",
    "h_marital", "h_parity", "cohort"
  ))) %>%
  filter(!is.na(cohort)) %>%
  group_by(cohort)

# split group by cohort
daf_impcheck_cohort <- group_split(daf_impcheck)
cohort <- group_keys(daf_impcheck)

# Proportion of missing values
# no cohort with complete cases
prop_missing <- data.frame(matrix(nrow = length(cohort), ncol = 3))
colnames(prop_missing) <- c("cohort", "#incompl.", "%incompl.")

for (i in 1:nrow(cohort)) {
  prop_missing[i, 1] <- as.character(cohort[i, ]$cohort)
  prop_missing[i, 2] <- table(ifelse(complete.cases(daf_impcheck_cohort[[i]]), "complete", "incompl."))
  prop_missing[i, 3] <- round(100 * table(complete.cases(daf_impcheck_cohort[[i]])) / nrow(daf_impcheck_cohort[[i]]), 2)
}

kable(prop_missing) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "responsive"), full_width = F) %>%
  kable_classic(full_width = F, html_font = "Cambria")

prop_missing_pervar <- as.data.frame(cbind(
  "# NA-SAB" = colSums(is.na(daf_impcheck_cohort[[1]])),
  "% NA-SAB" = round(colMeans(is.na(daf_impcheck_cohort[[1]])) * 100, 2),
  "# NA-EDEN" = colSums(is.na(daf_impcheck_cohort[[2]])),
  "% NA-EDEN" = round(colMeans(is.na(daf_impcheck_cohort[[2]])) * 100, 2),
  "# NA-BIB" = colSums(is.na(daf_impcheck_cohort[[3]])),
  "% NA-BIB" = round(colMeans(is.na(daf_impcheck_cohort[[3]])) * 100, 2),
  "# NA-RHEA" = colSums(is.na(daf_impcheck_cohort[[4]])),
  "% NA-RHEA" = round(colMeans(is.na(daf_impcheck_cohort[[4]])) * 100, 2),
  "# NA-KANC" = colSums(is.na(daf_impcheck_cohort[[5]])),
  "% NA-KANC" = round(colMeans(is.na(daf_impcheck_cohort[[5]])) * 100, 2),
  "# NA-MOBA" = colSums(is.na(daf_impcheck_cohort[[6]])),
  "% NA-MOBA" = round(colMeans(is.na(daf_impcheck_cohort[[6]])) * 100, 2)
))

prop_missing_pervar %>%
  filter(row_number() <= n() - 1) %>% # remove last row: cohort
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")

# write.table(tbl_missing, file = 'proportion of missings_1y.txt', sep = ',', quote = F, row.names = F)
```

# Descriptive analysis of biomarkers

The following descriptive statistics were estimated based on complete cases.

```{r descriptive statistic func}
continuous_stats <- function(data, var) {
  data %>%
    group_by(cohort) %>%
    summarise(
      variable = quo_name(var),
      n = sum(!is.na({{ var }})),
      min = min({{ var }}, na.rm = TRUE),
      q1 = quantile({{ var }}, probs = 0.25, na.rm = TRUE),
      median = median({{ var }}, na.rm = TRUE),
      q3 = quantile({{ var }}, probs = 0.75, na.rm = TRUE),
      max = max({{ var }}, na.rm = TRUE),
      mean = mean({{ var }}, na.rm = TRUE),
      sd = sd({{ var }}, na.rm = TRUE),
    )
}

# interesting function, calculate quantiles
# quibble2 <- function(x, q = c(0.25, 0.5, 0.75)) {
#  tibble("{{ x }}" := quantile(x, q, na.rm = T), "{{ x }}_q" := q)
# }
```

## Cardiovascular biomarkers

### Missing percentage

```{r missing cardio}
daf_impcheck <- df_final %>%
  select(cardio, cohort) %>%
  filter(!is.na(cohort)) %>%
  group_by(cohort)

# split group by cohort
daf_impcheck_cohort <- group_split(daf_impcheck)
cohort <- group_keys(daf_impcheck)

prop_missing_pervar <- as.data.frame(cbind(
  "# NA-SAB" = colSums(is.na(daf_impcheck_cohort[[1]])),
  "% NA-SAB" = round(colMeans(is.na(daf_impcheck_cohort[[1]])) * 100, 2),
  "# NA-EDEN" = colSums(is.na(daf_impcheck_cohort[[2]])),
  "% NA-EDEN" = round(colMeans(is.na(daf_impcheck_cohort[[2]])) * 100, 2),
  "# NA-BIB" = colSums(is.na(daf_impcheck_cohort[[3]])),
  "% NA-BIB" = round(colMeans(is.na(daf_impcheck_cohort[[3]])) * 100, 2),
  "# NA-RHEA" = colSums(is.na(daf_impcheck_cohort[[4]])),
  "% NA-RHEA" = round(colMeans(is.na(daf_impcheck_cohort[[4]])) * 100, 2),
  "# NA-KANC" = colSums(is.na(daf_impcheck_cohort[[5]])),
  "% NA-KANC" = round(colMeans(is.na(daf_impcheck_cohort[[5]])) * 100, 2),
  "# NA-MOBA" = colSums(is.na(daf_impcheck_cohort[[6]])),
  "% NA-MOBA" = round(colMeans(is.na(daf_impcheck_cohort[[6]])) * 100, 2)
))

prop_missing_pervar %>%
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")
```

```{r cardio}
cardio_desc <- map_dfr(
  data = df_final,
  .x = quos(hs_zsys_bp, hs_zdia_bp, bp_pulse_avg),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "total")

cardio_desc_g <- map_dfr(
  data = df_finalg,
  .x = quos(hs_zsys_bp, hs_zdia_bp, bp_pulse_avg),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "girl")

cardio_desc_b <- map_dfr(
  data = df_finalb,
  .x = quos(hs_zsys_bp, hs_zdia_bp, bp_pulse_avg),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "boy")

cardio_desc_all <- rbind(cardio_desc, cardio_desc_g, cardio_desc_b) %>%
  arrange(match(group, c("total", "girl", "boy")), variable, cohort)

cardio_desc_all %>%
  kable(col.names = c(
    "Cohort", "Biomarkers",
    "N",
    "Min",
    "Q1",
    "Median",
    "Q3",
    "Max",
    "Mean",
    "Std",
    "Group"
  ), escape = F) %>%
  column_spec(11, bold = T) %>%
  row_spec(which(cardio_desc_all$group == "total"), color = "black", background = "#bdbdbd") %>%
  row_spec(which(cardio_desc_all$group == "girl"), color = "black", background = "white") %>%
  row_spec(which(cardio_desc_all$group == "boy"), color = "black", background = "#bdbdbd") %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "300px")
```

### by sex

```{r cardio_plot_by_sex}
cardio_df <- df_final %>%
  select(e3_sex, hs_zsys_bp, hs_zdia_bp, bp_pulse_avg)

cardio_df %>%
  pivot_longer(cols = 2:ncol(cardio_df)) %>%
  ggplot(aes(x = value, fill = e3_sex)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by age group

```{r cardio_plot_by_agegrp}
cardio_df <- df_final %>%
  select(child_age_grp, hs_zsys_bp, hs_zdia_bp, bp_pulse_avg)

cardio_df %>%
  pivot_longer(cols = 2:ncol(cardio_df)) %>%
  ggplot(aes(x = value, fill = child_age_grp)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by cohort

```{r cardio_plot_by_cohort}
cardio_df <- df_final %>%
  select(cohort, hs_zsys_bp, hs_zdia_bp, bp_pulse_sd)

cardio_df %>%
  pivot_longer(cols = 2:ncol(cardio_df)) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by maternal education

```{r cardio_plot_by_edum}
cardio_df <- df_final %>%
  select(h_edumc, hs_zsys_bp, hs_zdia_bp, bp_pulse_avg)

cardio_df %>%
  pivot_longer(cols = 2:ncol(cardio_df)) %>%
  ggplot(aes(x = value, fill = h_edumc)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by FAS (family affluent scale)

```{r cardio_plot_by_fas}
cardio_df <- df_final %>%
  select(FAS_cat, hs_zsys_bp, hs_zdia_bp, bp_pulse_avg)

cardio_df %>%
  pivot_longer(cols = 2:ncol(cardio_df)) %>%
  ggplot(aes(x = value, fill = FAS_cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by child ethnicity

```{r cardio_plot_by_ethc}
cardio_df <- df_final %>%
  select(h_ethnicity_3cat, hs_zsys_bp, hs_zdia_bp, bp_pulse_avg)

cardio_df %>%
  pivot_longer(cols = 2:ncol(cardio_df)) %>%
  ggplot(aes(x = value, fill = h_ethnicity_3cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

## Metabolic biomarkers

Unit z score

### Missing percentage

```{r missing metab}
daf_impcheck <- df_final %>%
  select(metab, cohort) %>%
  filter(!is.na(cohort)) %>%
  group_by(cohort)

# split group by cohort
daf_impcheck_cohort <- group_split(daf_impcheck)
cohort <- group_keys(daf_impcheck)

prop_missing_pervar <- as.data.frame(cbind(
  "# NA-SAB" = colSums(is.na(daf_impcheck_cohort[[1]])),
  "% NA-SAB" = round(colMeans(is.na(daf_impcheck_cohort[[1]])) * 100, 2),
  "# NA-EDEN" = colSums(is.na(daf_impcheck_cohort[[2]])),
  "% NA-EDEN" = round(colMeans(is.na(daf_impcheck_cohort[[2]])) * 100, 2),
  "# NA-BIB" = colSums(is.na(daf_impcheck_cohort[[3]])),
  "% NA-BIB" = round(colMeans(is.na(daf_impcheck_cohort[[3]])) * 100, 2),
  "# NA-RHEA" = colSums(is.na(daf_impcheck_cohort[[4]])),
  "% NA-RHEA" = round(colMeans(is.na(daf_impcheck_cohort[[4]])) * 100, 2),
  "# NA-KANC" = colSums(is.na(daf_impcheck_cohort[[5]])),
  "% NA-KANC" = round(colMeans(is.na(daf_impcheck_cohort[[5]])) * 100, 2),
  "# NA-MOBA" = colSums(is.na(daf_impcheck_cohort[[6]])),
  "% NA-MOBA" = round(colMeans(is.na(daf_impcheck_cohort[[6]])) * 100, 2)
))

prop_missing_pervar %>%
  filter(row_number() <= n() - 1) %>% # remove last row: cohort
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")
```

### Anthrpometrics

```{r anthropometrics}
anthro_desc <- map_dfr(
  data = df_final,
  .x = quos(hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "total")

anthro_desc_g <- map_dfr(
  data = df_finalg,
  .x = quos(hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "girl")

anthro_desc_b <- map_dfr(
  data = df_finalb,
  .x = quos(hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "boy")

anthro_desc_all <- rbind(anthro_desc, anthro_desc_g, anthro_desc_b) %>%
  arrange(match(group, c("total", "girl", "boy")), variable, cohort)

anthro_desc_all %>%
  kable(col.names = c(
    "Cohort", "Biomarkers",
    "N",
    "Min",
    "Q1",
    "Median",
    "Q3",
    "Max",
    "Mean",
    "Std",
    "Group"
  ), escape = F) %>%
  column_spec(11, bold = T) %>%
  row_spec(which(anthro_desc_all$group == "total"), color = "black", background = "#bdbdbd") %>%
  row_spec(which(anthro_desc_all$group == "girl"), color = "black", background = "white") %>%
  row_spec(which(anthro_desc_all$group == "boy"), color = "black", background = "#bdbdbd") %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "300px")
```

#### by sex

```{r anthropometrics_plot_by_sex}
anthro_df <- df_final %>%
  select(e3_sex, hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia)
anthro_df %>%
  pivot_longer(cols = 2:ncol(anthro_df)) %>%
  ggplot(aes(x = value, fill = e3_sex)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by age group

```{r anthropometrics_plot_by_agegrp}
anthro_df <- df_final %>%
  select(child_age_grp, hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia)
anthro_df %>%
  pivot_longer(cols = 2:ncol(anthro_df)) %>%
  ggplot(aes(x = value, fill = child_age_grp)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by cohort

```{r anthropometrics_plot_by_cohort}
anthro_df <- df_final %>%
  select(cohort, hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia)
anthro_df %>%
  pivot_longer(cols = 2:ncol(anthro_df)) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by maternal education

```{r anthropometrics_plot_by_edum}
anthro_df <- df_final %>%
  select(h_edumc, hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia)
anthro_df %>%
  pivot_longer(cols = 2:ncol(anthro_df)) %>%
  ggplot(aes(x = value, fill = h_edumc)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by FAS

```{r anthropometrics_plot_by_fas}
anthro_df <- df_final %>%
  select(FAS_cat, hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia)
anthro_df %>%
  pivot_longer(cols = 2:ncol(anthro_df)) %>%
  ggplot(aes(x = value, fill = FAS_cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by child ethnicity

```{r anthropometrics_plot_by_ethc}
anthro_df <- df_final %>%
  select(h_ethnicity_3cat, hs_zbmi_who, hs_zwaist_mets, hs_skf_sum2, hs_freefatmass_bia, hs_fatmass_bia, hs_fatprop_bia)
anthro_df %>%
  pivot_longer(cols = 2:ncol(anthro_df)) %>%
  ggplot(aes(x = value, fill = h_ethnicity_3cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

Unit mg/dL

```{r metabolic}
metab_desc <- map_dfr(
  data = df_final,
  .x = quos(
    hs_hdlchol_c, hs_ldlchol_c, hs_triglyc_c, hs_totchol_c, Leptin, Adiponectin,
    CRP, INSULIN, Glucose, hs_creatinine_c
  ),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "total")

metab_desc_g <- map_dfr(
  data = df_finalg,
  .x = quos(
    hs_hdlchol_c, hs_ldlchol_c, hs_triglyc_c, hs_totchol_c, Leptin, Adiponectin,
    CRP, INSULIN, Glucose, hs_creatinine_c
  ),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "girl")

metab_desc_b <- map_dfr(
  data = df_finalb,
  .x = quos(
    hs_hdlchol_c, hs_ldlchol_c, hs_triglyc_c, hs_totchol_c, Leptin, Adiponectin,
    CRP, INSULIN, Glucose, hs_creatinine_c
  ),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "boy")

metab_desc_all <- rbind(metab_desc, metab_desc_g, metab_desc_b) %>%
  arrange(match(group, c("total", "girl", "boy")), variable, cohort)

metab_desc_all %>%
  kable(col.names = c(
    "Cohort",
    "Biomarkers",
    "N",
    "Min",
    "Q1",
    "Median",
    "Q3",
    "Max",
    "Mean",
    "Std",
    "Group"
  ), escape = F) %>%
  column_spec(11, bold = T) %>%
  row_spec(which(metab_desc_all$group == "total"), color = "black", background = "#bdbdbd") %>%
  row_spec(which(metab_desc_all$group == "girl"), color = "black", background = "white") %>%
  row_spec(which(metab_desc_all$group == "boy"), color = "black", background = "#bdbdbd") %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "300px")
```

### Metabolic biomarkers

#### by sex

```{r metabolic_plot_by_sex}
metab_df <- df_final %>%
  select(
    e3_sex, hs_hdlchol_c, hs_non_hdl_c, hs_totchol_c, hs_ldlchol_c, hs_triglyc_c, Leptin, Adiponectin,
    CRP, INSULIN, Glucose, hs_creatinine_c
  )
metab_df %>%
  pivot_longer(cols = 2:ncol(metab_df)) %>%
  ggplot(aes(x = value, fill = e3_sex)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by age group

```{r metabolic_plot_by_agegrp}
metab_df <- df_final %>%
  select(
    child_age_grp, hs_hdlchol_c, hs_non_hdl_c, hs_totchol_c, hs_ldlchol_c, hs_triglyc_c, Leptin, Adiponectin, CRP, INSULIN, Glucose, hs_creatinine_c
  )
metab_df %>%
  pivot_longer(cols = 2:ncol(metab_df)) %>%
  ggplot(aes(x = value, fill = child_age_grp)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by cohort

```{r metabolic_plot_by_cohort}
metab_df <- df_final %>%
  select(
    cohort, hs_hdlchol_c, hs_non_hdl_c, hs_totchol_c, hs_ldlchol_c, hs_triglyc_c, Leptin, Adiponectin, CRP, INSULIN, Glucose, hs_creatinine_c
  )
metab_df %>%
  pivot_longer(cols = 2:ncol(metab_df)) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by maternal education

```{r metabolic_plot_by_edum}
metab_df <- df_final %>%
  select(
    h_edumc, hs_hdlchol_c, hs_non_hdl_c, hs_totchol_c, hs_ldlchol_c, hs_triglyc_c, Leptin, Adiponectin, CRP, INSULIN, Glucose, hs_creatinine_c
  )
metab_df %>%
  pivot_longer(cols = 2:ncol(metab_df)) %>%
  ggplot(aes(x = value, fill = h_edumc)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by FAS

```{r metabolic_plot_by_fas}
metab_df <- df_final %>%
  select(
    FAS_cat, hs_hdlchol_c, hs_non_hdl_c, hs_totchol_c, hs_ldlchol_c, hs_triglyc_c, Leptin, Adiponectin, CRP, INSULIN, Glucose, hs_creatinine_c
  )
metab_df %>%
  pivot_longer(cols = 2:ncol(metab_df)) %>%
  ggplot(aes(x = value, fill = FAS_cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

#### by child ethnicity

```{r metabolic_plot_by_ethc}
metab_df <- df_final %>%
  select(
    h_ethnicity_3cat, hs_hdlchol_c, hs_non_hdl_c, hs_totchol_c, hs_ldlchol_c, hs_triglyc_c, Leptin, Adiponectin, CRP, INSULIN, Glucose, hs_creatinine_c
  )
metab_df %>%
  pivot_longer(cols = 2:ncol(metab_df)) %>%
  ggplot(aes(x = value, fill = h_ethnicity_3cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

## Inflammatory biomarkers

Unit pg/mL, log2 transformed, imputed and normalized

### Missing percentage

```{r missing immune}
daf_impcheck <- df_final %>%
  select(immune, cohort) %>%
  filter(!is.na(cohort)) %>%
  group_by(cohort)

# split group by cohort
daf_impcheck_cohort <- group_split(daf_impcheck)
cohort <- group_keys(daf_impcheck)

prop_missing_pervar <- as.data.frame(cbind(
  "# NA-SAB" = colSums(is.na(daf_impcheck_cohort[[1]])),
  "% NA-SAB" = round(colMeans(is.na(daf_impcheck_cohort[[1]])) * 100, 2),
  "# NA-EDEN" = colSums(is.na(daf_impcheck_cohort[[2]])),
  "% NA-EDEN" = round(colMeans(is.na(daf_impcheck_cohort[[2]])) * 100, 2),
  "# NA-BIB" = colSums(is.na(daf_impcheck_cohort[[3]])),
  "% NA-BIB" = round(colMeans(is.na(daf_impcheck_cohort[[3]])) * 100, 2),
  "# NA-RHEA" = colSums(is.na(daf_impcheck_cohort[[4]])),
  "% NA-RHEA" = round(colMeans(is.na(daf_impcheck_cohort[[4]])) * 100, 2),
  "# NA-KANC" = colSums(is.na(daf_impcheck_cohort[[5]])),
  "% NA-KANC" = round(colMeans(is.na(daf_impcheck_cohort[[5]])) * 100, 2),
  "# NA-MOBA" = colSums(is.na(daf_impcheck_cohort[[6]])),
  "% NA-MOBA" = round(colMeans(is.na(daf_impcheck_cohort[[6]])) * 100, 2)
))

prop_missing_pervar %>%
  filter(row_number() <= n() - 1) %>% # remove last row: cohort
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")
```

```{r immune}
immune_desc <- map_dfr(
  data = df_final,
  .x = quos(CRP, IL1beta, IL6, IL8, IL10, TNFalfa),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "total")

immune_desc_g <- map_dfr(
  data = df_finalg,
  .x = quos(CRP, IL1beta, IL6, IL8, IL10, TNFalfa),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "girl")

immune_desc_b <- map_dfr(
  data = df_finalb,
  .x = quos(CRP, IL1beta, IL6, IL8, IL10, TNFalfa),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "boy")

immune_desc_all <- rbind(immune_desc, immune_desc_g, immune_desc_b) %>%
  arrange(match(group, c("total", "girl", "boy")), variable, cohort)

immune_desc_all %>%
  kable(col.names = c(
    "Cohort", "Biomarkers",
    "N",
    "Min",
    "Q1",
    "Median",
    "Q3",
    "Max",
    "Mean",
    "Std",
    "Group"
  ), escape = F) %>%
  column_spec(11, bold = T) %>%
  row_spec(which(immune_desc_all$group == "total"), color = "black", background = "#bdbdbd") %>%
  row_spec(which(immune_desc_all$group == "girl"), color = "black", background = "white") %>%
  row_spec(which(immune_desc_all$group == "boy"), color = "black", background = "#bdbdbd") %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "300px")
```

### by sex

```{r immune_plot_by_sex}
immune_df <- df_final %>%
  select(e3_sex, CRP, IL1beta, IL6, IL8, IL10, TNFalfa)
immune_df %>%
  pivot_longer(cols = 2:ncol(immune_df)) %>%
  ggplot(aes(x = value, fill = e3_sex)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by age group

```{r immune_plot_by_agegrp}
immune_df <- df_final %>%
  select(child_age_grp, CRP, IL1beta, IL6, IL8, IL10, TNFalfa)
immune_df %>%
  pivot_longer(cols = 2:ncol(immune_df)) %>%
  ggplot(aes(x = value, fill = child_age_grp)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by cohort

```{r immune_plot_by_cohort}
immune_df <- df_final %>%
  select(cohort, CRP, IL1beta, IL6, IL8, IL10, TNFalfa)
immune_df %>%
  pivot_longer(cols = 2:ncol(immune_df)) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by maternal education

```{r immune_plot_by_edum}
immune_df <- df_final %>%
  select(h_edumc, CRP, IL1beta, IL6, IL8, IL10, TNFalfa)
immune_df %>%
  pivot_longer(cols = 2:ncol(immune_df)) %>%
  ggplot(aes(x = value, fill = h_edumc)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by FAS

```{r immune_plot_by_fas}
immune_df <- df_final %>%
  select(FAS_cat, CRP, IL1beta, IL6, IL8, IL10, TNFalfa)
immune_df %>%
  pivot_longer(cols = 2:ncol(immune_df)) %>%
  ggplot(aes(x = value, fill = FAS_cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by child ethnicity

```{r immune_plot_by_ethc}
immune_df <- df_final %>%
  select(h_ethnicity_3cat, CRP, IL1beta, IL6, IL8, IL10, TNFalfa)
immune_df %>%
  pivot_longer(cols = 2:ncol(immune_df)) %>%
  ggplot(aes(x = value, fill = h_ethnicity_3cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

## Neuroendocrine biomarkers

### Missing percentage

```{r missing neuro}
daf_impcheck <- df_final %>%
  select(neuro, cohort) %>%
  filter(!is.na(cohort)) %>%
  group_by(cohort)

# split group by cohort
daf_impcheck_cohort <- group_split(daf_impcheck)
cohort <- group_keys(daf_impcheck)

prop_missing_pervar <- as.data.frame(cbind(
  "# NA-SAB" = colSums(is.na(daf_impcheck_cohort[[1]])),
  "% NA-SAB" = round(colMeans(is.na(daf_impcheck_cohort[[1]])) * 100, 2),
  "# NA-EDEN" = colSums(is.na(daf_impcheck_cohort[[2]])),
  "% NA-EDEN" = round(colMeans(is.na(daf_impcheck_cohort[[2]])) * 100, 2),
  "# NA-BIB" = colSums(is.na(daf_impcheck_cohort[[3]])),
  "% NA-BIB" = round(colMeans(is.na(daf_impcheck_cohort[[3]])) * 100, 2),
  "# NA-RHEA" = colSums(is.na(daf_impcheck_cohort[[4]])),
  "% NA-RHEA" = round(colMeans(is.na(daf_impcheck_cohort[[4]])) * 100, 2),
  "# NA-KANC" = colSums(is.na(daf_impcheck_cohort[[5]])),
  "% NA-KANC" = round(colMeans(is.na(daf_impcheck_cohort[[5]])) * 100, 2),
  "# NA-MOBA" = colSums(is.na(daf_impcheck_cohort[[6]])),
  "% NA-MOBA" = round(colMeans(is.na(daf_impcheck_cohort[[6]])) * 100, 2)
))

prop_missing_pervar %>%
  filter(row_number() <= n() - 1) %>% # remove last row: cohort
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")
```

```{r cortisol}
cort_desc <- map_dfr(
  data = df_final,
  .x = quos(cortisol_prod_log10),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "total")

cort_desc_g <- map_dfr(
  data = df_finalg,
  .x = quos(cortisol_prod_log10),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "girl")

cort_desc_b <- map_dfr(
  data = df_finalb,
  .x = quos(cortisol_prod_log10),
  .f = continuous_stats
) %>%
  mutate(across(where(is.numeric), round, 2)) %>%
  mutate(group = "boy")

cort_desc_all <- rbind(cort_desc, cort_desc_g, cort_desc_b) %>%
  arrange(match(group, c("total", "girl", "boy")), variable, cohort)

cort_desc_all %>%
  kable(col.names = c(
    "Cohort", "Biomarkers",
    "N",
    "Min",
    "Q1",
    "Median",
    "Q3",
    "Max",
    "Mean",
    "Std",
    "Group"
  ), escape = F) %>%
  column_spec(11, bold = T) %>%
  row_spec(which(cort_desc_all$group == "total"), color = "black", background = "#bdbdbd") %>%
  row_spec(which(cort_desc_all$group == "girl"), color = "black", background = "white") %>%
  row_spec(which(cort_desc_all$group == "boy"), color = "black", background = "#bdbdbd") %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "300px")
```

### by sex

```{r cortisol_plot_by_sex}
cort <- df_final %>%
  select(e3_sex, cortisol_prod_log10)
cort %>%
  pivot_longer(cols = 2:ncol(cort)) %>%
  ggplot(aes(x = value, fill = e3_sex)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by age group

```{r cortisol_plot_by_agegrp}
cort <- df_final %>%
  select(child_age_grp, cortisol_prod_log10)
cort %>%
  pivot_longer(cols = 2:ncol(cort)) %>%
  ggplot(aes(x = value, fill = child_age_grp)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by cohort

```{r cortisol_plot_by_cohort}
cort <- df_final %>%
  select(cohort, cortisol_prod_log10)
cort %>%
  pivot_longer(cols = 2:ncol(cort)) %>%
  ggplot(aes(x = value, fill = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by maternal education

```{r cortisol_plot_by_edum}
cort <- df_final %>%
  select(h_edumc, cortisol_prod_log10)
cort %>%
  pivot_longer(cols = 2:ncol(cort)) %>%
  ggplot(aes(x = value, fill = h_edumc)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by FAS

```{r cortisol_plot_by_fas}
cort <- df_final %>%
  select(FAS_cat, cortisol_prod_log10)
cort %>%
  pivot_longer(cols = 2:ncol(cort)) %>%
  ggplot(aes(x = value, fill = FAS_cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

### by child ethnicity

```{r cortisol_plot_by_ethc}
cort <- df_final %>%
  select(h_ethnicity_3cat, cortisol_prod_log10)
cort %>%
  pivot_longer(cols = 2:ncol(cort)) %>%
  ggplot(aes(x = value, fill = h_ethnicity_3cat)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  facet_wrap(~name, scales = "free_x")
```

# Allostatic load scores

## Complete cases of outcomes

```{r complete_outcome}
out_desc <- df_final %>%
  filter(!is.na(e3_sex)) %>%
  select(cohort, all_of(out)) %>%
  group_by(cohort)

as.data.frame(cbind(
  "#" = colSums(!is.na(out_desc))
)) %>%
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")

out_desc_cohort <- group_split(out_desc)
cohort <- group_keys(out_desc)

outcome_desc <- as.data.frame(cbind(
  "# SAB" = colSums(!is.na(out_desc_cohort[[1]])),
  "# EDEN" = colSums(!is.na(out_desc_cohort[[2]])),
  "# BIB" = colSums(!is.na(out_desc_cohort[[3]])),
  "# RHEA" = colSums(!is.na(out_desc_cohort[[4]])),
  "# KANC" = colSums(!is.na(out_desc_cohort[[5]])),
  "# MOBA" = colSums(!is.na(out_desc_cohort[[6]]))
))

outcome_desc %>%
  tibble::rownames_to_column(var = "Variables") %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), full_width = T, fixed_thead = T) %>%
  kable_classic(full_width = F, html_font = "Cambria")
```

## Distribution by cohort

```{r}
p1 <- df_final %>%
  ggplot(aes(x = AL_2_tot, fill = cohort)) +
  geom_bar(position = "dodge") +
  ggtitle("Cut-off based AL score by cohort")

p2 <- df_final %>%
  ggplot(aes(x = AL_z2_tot, fill = cohort, color = cohort)) +
  geom_histogram(position = "identity", alpha = 0.5) +
  ggtitle("Z-score based AL score by cohort")

wrap_plots(p1, p2, ncol = 1, nrow = 2)
```

### Mean (SD) Table
```{r}
mean_sd <- list(
  mean = ~mean(.x, na.rm = T),
  sd = ~sd(.x, na.rm = T)
)
  
# Total score
AL_desc <- df_final %>%
  select(cohort, AL_2_tot, AL_z2_tot) %>% 
  group_by(cohort) %>% 
  summarise(across(c(AL_2_tot, AL_z2_tot), mean_sd)) %>% 
  mutate_at(2:5, round, 1) %>% 
  mutate(
    countAL = paste(AL_2_tot_mean, " (", AL_2_tot_sd, ")", sep = ""),
    continuousAL = paste(AL_z2_tot_mean, " (", AL_z2_tot_sd, ")", sep = "")
  ) %>% 
  select(cohort, countAL, continuousAL) %>% 
  t()


AL_desc %>%
  kable(
    caption = "Descriptives of allostatic load score by cohort",
    escape = F
  ) %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, position = "left", font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "100%")
```

### Median (IQR) Table
```{r}
median_iqr <- list(
  median = ~median(.x, na.rm = T),
  iqr25 = ~quantile(.x, na.rm = T)[[2]],
  iqr75 = ~quantile(.x, na.rm = T)[[4]]
)
  
# Total score
AL_desc <- df_final %>%
  select(cohort, AL_2_tot, AL_z2_tot) %>% 
  group_by(cohort) %>% 
  summarise(across(c(AL_2_tot, AL_z2_tot), median_iqr)) %>% 
  mutate_at(2:7, round, 1) %>% 
  mutate(
    countAL = paste(AL_2_tot_median, " (", AL_2_tot_iqr25, ", ", AL_2_tot_iqr75, ")", sep = ""),
    continuousAL = paste(AL_z2_tot_median, " (", AL_z2_tot_iqr25, ", ", AL_z2_tot_iqr75, ")", sep = "")
  ) %>% 
  select(cohort, countAL, continuousAL) %>% 
  t()


AL_desc %>%
  kable(
    caption = "Descriptives of allostatic load score by cohort",
    escape = F
  ) %>%
  kable_styling(bootstrap_options = c("condensed", "responsive"), full_width = F, position = "left", font_size = 14, fixed_thead = T) %>%
  scroll_box(width = "100%", height = "100%")
```

# Correlation matrix

## Air pollutants, noise level, covariates and allostatic load score

```{r corr_matrix_with_SES, fig.width = 18 , fig.height = 18}
df_final %>%
  select(
    expo, cov, out
  ) %>%
  mutate_if(is.factor, as.numeric) %>%
  cor(use = "pairwise.complete.obs", method = "pearson") %>%
  ggcorrplot(
    hc.order = TRUE,
    type = "lower",
    lab = TRUE
  )
```

## Correlations among allostatic load score

```{r corr_al, fig.width=10, fig.height=10}
df_final %>%
  select(all_of(out)) %>%
  mutate_if(is.factor, as.numeric) %>%
  cor(use = "pairwise.complete.obs", method = "pearson") %>%
  ggcorrplot(
    hc.order = TRUE,
    type = "lower",
    lab = TRUE
  )
```

## Correlation between biomarkers

```{r corr_biomarkers, fig.width=10, fig.height=10}
df_final %>%
  select(all_of(cardio), all_of(metab), all_of(immune), all_of(neuro)) %>%
  mutate_if(is.factor, as.numeric) %>%
  cor(use = "pairwise.complete.obs", method = "pearson") %>%
  ggcorrplot(
    hc.order = TRUE,
    type = "lower",
    lab = TRUE
  )

ggsave("/Users/ymou/helix_project/results/corr_biomarker.png", width = 10, height = 10, units = "in", dpi = 300)
```

Biomarkers with correlation estimates higher than 0.8 will be inspected IL-6 and IL-1b are highly correlated. BMI and weight circumferences are highly correlated.

```{r}
df_final %>%
  select(cohort, IL1beta, IL6, hs_zbmi_who, hs_zwaist_mets) %>%
  group_by(cohort) %>%
  summarise(
    IL_cor = cor(IL1beta, IL6, use = "complete.obs"),
    anthropometric_cor = cor(hs_zbmi_who, hs_zwaist_mets, use = "complete.obs")
  )
```

## Correlation between noise level at home and at school, categorical variable

```{r corr_noise_cat}
library(polycor)

table(df_final$hs_lden_c_h, df_final$hs_lden_c_s)

# calculate polychoric correlation
polychor(df_final$hs_lden_c_h, df_final$hs_lden_c_s, ML = T, std.err = T)
```

```{r output, include=FALSE, eval = F}
# convert rmd to r file
knitr::purl("/Users/ymou/helix_project/code/helix-allostasis/3a_descriptive_analysis_all_raw_data.Rmd", documentation = 1L)
```
