==> 5_flow_chart.Rmd <==
---
title: 'Flow chart: Air pollution, noise level and allostatic load'
author: "Yuchan Mou"
date: "`r Sys.Date()`"
output: 
  html_document:
editor_options: 
  chunk_output_type: console
---

```{r set_up, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

```{r load packages, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
# load pkgs
packages <- c("tidyverse", "Gmisc", "grid", "ggplot2", "glue")
sapply(packages, library, character.only = T)
```

```{r import data, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
load("/Users/ymou/helix_project/data/analysis_data/df_al_final_manipulated.RData")
Sys.Date()
```

### Flow chart of the study population selection
Complete cases of allostatic load scores
vignette: https://cran.r-project.org/web/packages/Gmisc/vignettes/Grid-based_flowcharts.html
```{r flowchart, fig.width = 10 , fig.height = 10}
# Set dimensions for 300 DPI
width_in_inches <- 12
height_in_inches <- 15
dpi <- 300

# Calculate dimensions in pixels
width_px <- width_in_inches * dpi
height_px <- height_in_inches * dpi

# Open PNG device to save the flowchart
png("/Users/ymou/helix_project/results/Fig/flowchart.png", width = width_px, height = height_px, res = dpi)

# Create the flowchart
grid.newpage()
# set parameters
midx <- 0.5
x1 <- 0.1
x2 <- 0.26
x2_5 <- 0.34
x3 <- 0.42
x4 <- 0.58
x5 <- 0.74
x6 <- 0.90
width <- 0.16
gp <- gpar(fill = "white")

# create boxes
(org_cohort <- boxGrob(
  glue("Mother and child pairs in the HELIX subcohort",
    "n = {pop}",
    pop = nrow(df_final),
    .sep = "\n"
  ),
  x = midx, y = .9, box_gp = gp, width = 0.4
))

(BIB <- boxGrob(
  str_glue("BIB (N = {pop})\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    pop = length(which(df_final$cohort == "BIB")),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_no2_yr_hs_t))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_pm25_yr_hs_t))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_pm25abs_yr_hs_t))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_pm10_yr_hs_t))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_lden_c_h))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_lden_c_s)))
  ),
  x = x1, y = .7, box_gp = gp, width = width, just = "left"
))

(BIB_1 <- boxGrob(
  str_glue("With all biomarkers\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    #min = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot))),
    #max = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "BIB") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x1, y = .5, box_gp = gp, width = width, just = "left"
))

connectGrob(org_cohort, BIB, type = "N", "l")
connectGrob(BIB, BIB_1, type = "N", "l")


(EDEN <- boxGrob(
  str_glue("EDEN (N = {pop})\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    pop = length(which(df_final$cohort == "EDEN")),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_no2_yr_hs_t))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_pm25_yr_hs_t))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_pm25abs_yr_hs_t))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_pm10_yr_hs_t))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_h))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_s)))
  ),
  x = x2, y = .7, box_gp = gp, width = width, just = "left"
))

(EDEN_1 <- boxGrob(
  str_glue("With all biomarkers\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    #min = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    #max = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x2, y = .5, box_gp = gp, width = width, just = "left"
))

(EDEN_2 <- boxGrob(
  str_glue("Exclude:\n-missing>70%\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "EDEN") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x2_5, y = .37, box_gp = gp, width = 0.14, just = "left"
))

connectGrob(org_cohort, EDEN, type = "N", "l")
connectGrob(EDEN, EDEN_1, type = "N", "l")
connectGrob(EDEN_1, EDEN_2, type = "L", "l")


(INMA <- boxGrob(
  str_glue("INMA (N = {pop})\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    pop = length(which(df_final$cohort == "SAB")),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_no2_yr_hs_t))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_pm25_yr_hs_t))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_pm25abs_yr_hs_t))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_pm10_yr_hs_t))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_lden_c_h))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_lden_c_s)))
  ),
  x = x3, y = .7, box_gp = gp, width = width, just = "left"
))

# missing patterns are the same across air pollutants
# for lden, home sample sizes are larger than school
# min sample size = lden_school, max sample size = no2
(INMA_1 <- boxGrob(
  str_glue("With all biomarkers\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    #min = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot))),
    #max = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "SAB") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x3, y = .5, box_gp = gp, width = width, just = "left"
))

connectGrob(org_cohort, INMA, type = "N", "l")
connectGrob(INMA, INMA_1, type = "N", "l")


(KANC <- boxGrob(
  str_glue("KANC (N = {pop})\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    pop = length(which(df_final$cohort == "KANC")),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_no2_yr_hs_t))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_pm25_yr_hs_t))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_pm25abs_yr_hs_t))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_pm10_yr_hs_t))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_lden_c_h))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_lden_c_s)))
  ),
  x = x4, y = .7, box_gp = gp, width = width, just = "left"
))

(KANC_1 <- boxGrob(
  str_glue("With all biomarkers\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    #min = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot))),
    #max = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "KANC") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x4, y = .5, box_gp = gp, width = width, just = "left"
))

connectGrob(org_cohort, KANC, type = "N", "l")
connectGrob(KANC, KANC_1, type = "N", "l")


(MOBA <- boxGrob(
  str_glue("MOBA (N = {pop})\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    pop = length(which(df_final$cohort == "MOBA")),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_no2_yr_hs_t))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_pm25_yr_hs_t))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_pm25abs_yr_hs_t))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_pm10_yr_hs_t))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_lden_c_h))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_lden_c_s)))
  ),
  x = x5, y = .7, box_gp = gp, width = width, just = "left"
))

(MOBA_1 <- boxGrob(
  str_glue("With all biomarkers\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    #min = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot))),
    #max = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "MOBA") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x5, y = .5, box_gp = gp, width = width, just = "left"
))

connectGrob(org_cohort, MOBA, type = "N", "l")
connectGrob(MOBA, MOBA_1, type = "N", "l")


(RHEA <- boxGrob(
  str_glue("RHEA (N = {pop})\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    pop = length(which(df_final$cohort == "RHEA")),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_no2_yr_hs_t))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_pm25_yr_hs_t))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_pm25abs_yr_hs_t))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_pm10_yr_hs_t))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_lden_c_h))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_lden_c_s)))
  ),
  x = x6, y = .7, box_gp = gp, width = width, just = "left"
))

(RHEA_1 <- boxGrob(
  str_glue("With all biomarkers\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    #min = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot))),
    #max = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_no2_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort == "RHEA") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot)))
  ),
  x = x6, y = .5, box_gp = gp, width = width, just = "left"
))

connectGrob(org_cohort, RHEA, type = "N", "l")
connectGrob(RHEA, RHEA_1, type = "N", "l")

# total study sample
(tot <- boxGrob(
  str_glue("Total study sample\n n = {min} ~ {max}\nNO2: n = {hs_no2_yr_hs_t}\nPM2.5: n = {hs_pm25_yr_hs_t}\nPM2.5abs: n= {hs_pm25abs_yr_hs_t}\nPM10: n = {hs_pm10_yr_hs_t}\nLden home: n = {hs_lden_c_h}\nLden school: n = {hs_lden_c_s}",
    hs_no2_yr_hs_t = nrow(df_final %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25_yr_hs_t = nrow(df_final %>% filter(!is.na(hs_pm25_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm25abs_yr_hs_t = nrow(df_final %>% filter(!is.na(hs_pm25abs_yr_hs_t) & !is.na(AL_2_tot))),
    hs_pm10_yr_hs_t = nrow(df_final %>% filter(!is.na(hs_pm10_yr_hs_t) & !is.na(AL_2_tot))),
    hs_lden_c_h = nrow(df_final %>% filter(cohort != "EDEN") %>% filter(!is.na(hs_lden_c_h) & !is.na(AL_2_tot))),
    hs_lden_c_s = nrow(df_final %>% filter(cohort != "EDEN") %>% filter(!is.na(hs_lden_c_s) & !is.na(AL_2_tot))),
    min = hs_pm25abs_yr_hs_t,
    max = nrow(df_final %>% filter(!is.na(hs_no2_yr_hs_t) & !is.na(AL_2_tot)))
  ),
  x = midx, y = .14, box_gp = gp, width = width, just = "left"
))

connectGrob(INMA_1, tot, type = "N", "l")
connectGrob(EDEN_1, tot, type = "N", "l")
connectGrob(BIB_1, tot, type = "N", "l")
connectGrob(RHEA_1, tot, type = "N", "l")
connectGrob(KANC_1, tot, type = "N", "l")
connectGrob(MOBA_1, tot, type = "N", "l")


# Close the PNG device
dev.off()
```
