==> 6_imputation_full.Rmd <==
---
title: 'Data imputation: Air pollution, noise level and allostatic load'
author: "Yuchan Mou"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
options(knitr.table.format = "html")
```

```{r load packages, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
# load pkgs
packages <- c("tidyverse", "mice", "miceadds", "plyr", "labelled")
sapply(packages, library, character.only = T)
```

```{r import data, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
load("/Users/ymou/helix_project/data/analysis_data/df_al_final_manipulated.RData")
Sys.Date()
```

```{r define variables, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
ap <- c(
  "hs_no2_yr_hs_t",
  "hs_pm25_yr_hs_t",
  "hs_pm25abs_yr_hs_t",
  "hs_pm10_yr_hs_t"
)
ap_supl <- c(
  "hs_no2_yr_hs_h", "hs_no2_yr_hs_s", "hs_no2_yr_hs_r", "hs_no2_yr_hs_p",
  "hs_pm25_yr_hs_h", "hs_pm25_yr_hs_s", "hs_pm25_yr_hs_r", "hs_pm25_yr_hs_p",
  "hs_pm25abs_yr_hs_h", "hs_pm25abs_yr_hs_s", "hs_pm25abs_yr_hs_r", "hs_pm25abs_yr_hs_p",
  "hs_pm10_yr_hs_h", "hs_pm10_yr_hs_s", "hs_pm10_yr_hs_r", "hs_pm10_yr_hs_p"
)
noise <- c("hs_lden_c_h", "hs_lden_c_s")
noise_supl <- c("hs_lden_h", "hs_lden_s", "hs_lden_t")

cov <- c(
  "e3_sex", "hs_mvpa", "hs_sd_wk", "hs_child_age_years",
  "FAS_cat", "h_native_2cat", "h_ethnicity_cauc", "h_edumc", "h_edufc",
  "hs_smk_parents", "hs_globalexp_2cat", "e3_asmokyn_p", "e3_alcpreg_yn", "h_marital_2cat", "h_age", "h_parity"
)

out <- c("AL_2_tot", "AL_z2_tot")
out_supl <- c(
              "AL_2_cardio", "AL_2_metab", "AL_2_immune", "AL_2_neuro",
              "AL_2_tot_wavg", "AL_2_cardio_wavg", "AL_2_metab_wavg", "AL_2_immune_wavg", "AL_2_neuro_wavg",
              "AL_z2_cardio", "AL_z2_metab", "AL_z2_immune", "AL_z2_neuro",
              "AL_z2_tot_wavg", "AL_z2_cardio_wavg", "AL_z2_metab_wavg", "AL_z2_immune_wavg", "AL_z2_neuro_wavg")
#"AL_1_tot", "AL_1_cardio", "AL_1_metab", "AL_1_immune", "AL_1_neuro",
#"AL_1_tot_wavg", "AL_1_cardio_wavg", "AL_1_metab_wavg", "AL_1_immune_wavg", "AL_1_neuro_wavg",
#"AL_z1_tot", "AL_z1_cardio", "AL_z1_metab", "AL_z1_immune", "AL_z1_neuro",
#"AL_z1_tot_wavg", "AL_z1_cardio_wavg", "AL_z1_metab_wavg", "AL_z1_immune_wavg", "AL_z1_neuro_wavg",

# 19 biomarkers included in the main AL score
# for imputation
biomarker_all <- c(
  "hs_zsys_bp", "hs_zdia_bp", "bp_pulse_sd",
  "hs_zbmi_who", "hs_zwaist_mets", "hs_skf_sum2", "hs_fatprop_bia",
  "hs_hdlchol_c", "hs_non_hdl_c", "hs_triglyc_c", "Leptin",
  "Adiponectin",
  "CRP", "IL1beta", "IL6", "IL8", "IL10", "TNFalfa",
  "cortisol_prod_log10"
)

cardio <- c("hs_zsys_bp", "hs_zdia_bp", "bp_pulse_sd") # 3 biomarkers
metab <- c(
  "hs_zbmi_who", "hs_zwaist_mets", "hs_skf_sum2", "hs_fatprop_bia",
  "hs_hdlchol_c", "hs_non_hdl_c", "hs_triglyc_c", "Leptin", "Adiponectin"
) # 9 biomarkers
immune <- c("CRP", "IL1beta", "IL6", "IL8", "IL10", "TNFalfa") # 6 biomarkers
neuro <- c("cortisol_prod_log10") # 1 biomarkers
```

```{r proportion plot for convergence check}
propplot <- function(x, formula, facet = "wrap", ...) {
  library(ggplot2)


  cd <- data.frame(mice::complete(x, "long", include = TRUE))
  cd$.imp <- factor(cd$.imp)

  r <- as.data.frame(is.na(x$data))

  impcat <- x$meth != "" & sapply(x$data, is.factor)
  vnames <- names(impcat)[impcat]

  if (missing(formula)) {
    formula <- as.formula(paste(paste(vnames,
      collapse = "+",
      sep = ""
    ), "~1", sep = ""))
  }

  tmsx <- terms(formula[-3], data = x$data)
  xnames <- attr(tmsx, "term.labels")
  xnames <- xnames[xnames %in% vnames]

  if (paste(formula[3]) != "1") {
    wvars <- gsub("[[:space:]]*\\|[[:print:]]*", "", paste(formula)[3])
    # wvars <- all.vars(as.formula(paste("~", wvars)))
    wvars <- attr(terms(as.formula(paste("~", wvars))), "term.labels")
    if (grepl("\\|", formula[3])) {
      svars <- gsub("[[:print:]]*\\|[[:space:]]*", "", paste(formula)[3])
      svars <- all.vars(as.formula(paste("~", svars)))
    } else {
      svars <- ".imp"
    }
  } else {
    wvars <- NULL
    svars <- ".imp"
  }

  for (i in seq_along(xnames)) {
    xvar <- xnames[i]
    select <- cd$.imp != 0 & !r[, xvar]
    cd[select, xvar] <- NA
  }


  for (i in which(!wvars %in% names(cd))) {
    cd[, wvars[i]] <- with(cd, eval(parse(text = wvars[i])))
  }

  meltDF <- reshape2::melt(cd[, c(wvars, svars, xnames)], id.vars = c(wvars, svars))
  meltDF <- meltDF[!is.na(meltDF$value), ]


  wvars <- if (!is.null(wvars)) paste0("`", wvars, "`")

  a <- plyr::ddply(meltDF, c(wvars, svars, "variable", "value"), plyr::summarize,
    count = length(value)
  )
  b <- plyr::ddply(meltDF, c(wvars, svars, "variable"), plyr::summarize,
    tot = length(value)
  )
  mdf <- merge(a, b)
  mdf$prop <- mdf$count / mdf$tot

  plotDF <- merge(unique(meltDF), mdf)
  plotDF$value <- factor(plotDF$value,
    levels = unique(unlist(lapply(x$data[, xnames], levels))),
    ordered = T
  )

  p <- ggplot(plotDF, aes(x = value, fill = get(svars), y = prop)) +
    geom_bar(position = "dodge", stat = "identity") +
    theme(legend.position = "bottom", ...) +
    ylab("proportion") +
    scale_fill_manual(
      name = "",
      values = c(
        "black",
        colorRampPalette(
          RColorBrewer::brewer.pal(9, "Blues")
        )(x$m + 3)[1:x$m + 3]
      )
    ) +
    guides(fill = guide_legend(nrow = 1))

  if (facet == "wrap") {
    if (length(xnames) > 1) {
      print(p + facet_wrap(c("variable", wvars), scales = "free"))
    } else {
      if (is.null(wvars)) {
        print(p)
      } else {
        print(p + facet_wrap(wvars, scales = "free"))
      }
    }
  }

  if (facet == "grid") {
    if (!is.null(wvars)) {
      print(p + facet_grid(paste(paste(wvars, collapse = "+"), "~ variable"),
        scales = "free"
      ))
    }
  }
}
```

# Add incl variable and split dataset
```{r incl_vars}
# Split cohort for imputation in each cohort separately
df_cohort <- list()
for (i in 1:6) {
  df_cohort[[i]] <- df_final %>% filter(cohort == levels(df_final$cohort)[i])
}

# Prepare the incl_vars to determine what should be imputed
for (i in 1:6) {
  df_cohort[[i]] <- df_cohort[[i]] %>%
    # calculate the number of biomarkers that are NA
    mutate(
      na_bio = rowSums(is.na(subset(df_cohort[[i]], select = biomarker_all)))
    ) %>%
    mutate(
      incl_var_ap = ifelse(!is.na(hs_no2_yr_hs_t) & !is.na(hs_pm25_yr_hs_t), 1, 0), # only these two are complete in every cohort
      incl_var_ap_all = ifelse(!is.na(hs_no2_yr_hs_t) & !is.na(hs_pm25_yr_hs_t) & !is.na(hs_pm25abs_yr_hs_t) & !is.na(hs_pm10_yr_hs_t), 1, 0),
      incl_var_noise = ifelse(!is.na(hs_lden_c_h) & !is.na(hs_lden_c_s), 1, 0),
      incl_var_out = ifelse(!is.na(AL_2_tot), 1, 0), # calculated AL score based on 100% biomarkers available
      incl_var_bio = ifelse(na_bio <= 2, 1, 0) # >90% biomarkers available are included
    )
}

# Dataset list filter by complete data in no2, pm2.5 and outcomes.
# EDEN is missing in PM2.5abs and PM10
# RHEA is missing in PM2.5abs
df_cohort_1 <- list()
df_cohort_imp <- list()
for (i in 1:6) {
  df_cohort_1[[i]] <- df_cohort[[i]] %>% 
    select(all_of(c("HelixID", ap, ap_supl, noise, noise_supl, cov, out, out_supl, biomarker_all, "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_out", "incl_var_bio", "cohort")), matches("_z$|_as1$")) 
  # complete NO2, PM2.5 with complete AL score or with 90% biomarkers available
  df_cohort_imp[[i]] <- subset(df_cohort_1[[i]], incl_var_ap == 1 & (incl_var_out == 1 | incl_var_bio == 1))
  if (length(df_cohort_1) == 6) {
    rm(df_cohort_1)
  }
}
```


# Imputation
RHEA has missing PM2.5abs
ap_supl, noise, noise_supl, out_supl, individual biomarker cutoff score and individual biomarker z score were not included in the imputation. They will be merged into the dataset after imputation. 

## SAB
```{r}
sab <- df_cohort_imp[[1]] %>% select(-all_of(c(ap_supl, noise, noise_supl, out_supl)), -matches("_z$|_as1$"))

inlist <- c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", `cov`[-c(7)], out)

imp_sab <- mice(sab,
  m = 25, max = 50, print = F,
  pred = quickpred(sab,
    mincor = 0.2, minpuc = 0, include = inlist,
    exclude = c(
      "HelixID", "hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t",
      "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_out", "incl_var_bio", "cohort", "h_ethnicity_cauc"
    )
  ),
  seed = 2023
)
# all air pollutants corrs >0.8
# all kids are caucasian. Therefore ethnicity_cauc is excl.

plot(imp_sab)
propplot(imp_sab, strip.text = element_text(size = 14))
```

## EDEN
EDEN has no data on PM2.5abs and PM10, and noise h/s.
```{r}
ede <- df_cohort_imp[[2]] %>% select(-all_of(c(
  "hs_pm25abs_yr_hs_t",
  "hs_pm10_yr_hs_t", ap_supl, noise, noise_supl, out_supl
)), -matches("_z$|_as1$"))

inlist <- c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", "hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t", `cov`[-c(5, 7)], out)

imp_ede <- mice(ede,
  m = 25, max = 50, print = F,
  pred = quickpred(ede,
    mincor = 0.2, minpuc = 0, include = inlist,
    exclude = c(
      "HelixID",
      "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_out", "incl_var_bio", "cohort", "h_ethnicity_cauc", "FAS_cat"
    )
  ),
  seed = 2023
)

# family income low is 0. Excluded in the imputation.

plot(imp_ede)
densityplot(imp_ede, ~hs_sd_wk)
propplot(imp_ede, strip.text = element_text(size = 14))
```

## BIB
```{r}
bib <- df_cohort_imp[[3]] %>% select(-all_of(c(ap_supl, noise, noise_supl, out_supl)), -matches("_z$|_as1$"))

inlist <- c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", "hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t", cov, out)

imp_bib <- mice(bib,
  m = 25, max = 50, print = F,
  pred = quickpred(bib,
    mincor = 0.2, minpuc = 0, include = inlist,
    exclude = c(
      "HelixID",
      "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_out", "incl_var_bio", "cohort"
    )
  ),
  seed = 2023
)

plot(imp_bib)
propplot(imp_bib, strip.text = element_text(size = 14))
```

## RHEA
RHEA has no data on PM2.5abs.
```{r}
rhea <- df_cohort_imp[[4]] %>% select(-all_of(c("hs_pm25abs_yr_hs_t", ap_supl, noise, noise_supl, out_supl)), -matches("_z$|_as1$"))

inlist <- c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", cov[-c(5, 7)], out)
# all air pollutants corrs >0.8
# force to include FAS will cause model convergence problem. therefore is not included in the inlist.

imp_rhea <- mice(rhea,
  m = 25, max = 50, print = F,
  pred = quickpred(rhea,
    mincor = 0.2, minpuc = 0, include = inlist,
    exclude = c(
      "HelixID", "hs_pm10_yr_hs_t",
      "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_bio", "incl_var_out", "cohort", "h_ethnicity_cauc"
    )
  ),
  seed = 2023
)
# all kids in rhea are caucasian. Therefore ethnicity_cauc was excl in imputation.

plot(imp_rhea)
propplot(imp_rhea, strip.text = element_text(size = 14))
```

## KANC
```{r}
kanc <- df_cohort_imp[[5]] %>% select(-all_of(c(ap_supl, noise, noise_supl, out_supl)), -matches("_z$|_as1$"))

inlist <- c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", "hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t", cov[-c(7)], out)

imp_kanc <- mice(kanc,
  m = 25, max = 50, print = F,
  pred = quickpred(kanc,
    mincor = 0.2, minpuc = 0, include = inlist,
    exclude = c(
      "HelixID",
      "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_out", "incl_var_bio", "cohort", "h_ethnicity_cauc"
    )
  ),
  seed = 2023
)

plot(imp_kanc)
propplot(imp_kanc, strip.text = element_text(size = 14))
```

## MOBA
```{r}
moba <- df_cohort_imp[[6]] %>% select(-all_of(c(ap_supl, noise, noise_supl, out_supl)), -matches("_z$|_as1$"))

inlist <- c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", "hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t", cov[-c(8, 9)], out)

imp_moba <- mice(moba,
  m = 25, max = 50, print = F,
  pred = quickpred(moba,
    mincor = 0.2, minpuc = 0, include = inlist,
    exclude = c(
      "HelixID",
      "incl_var_ap", "incl_var_ap_all", "incl_var_noise", "incl_var_bio", "incl_var_out", "cohort", "h_edumc", "h_edufc"
    )
  ),
  seed = 2023
)

plot(imp_moba)
propplot(imp_moba, strip.text = element_text(size = 14))
```

# Merge imputed dataset
```{r}
# set AL score where incl_var_out=0 as NA 
imp_sab_l <- complete(imp_sab, "long", include = T)
imp_sab_l <- imp_sab_l %>% 
  mutate(AL_2_tot = ifelse(imp_sab_l$incl_var_out == 1, AL_2_tot, NA),
         AL_z2_tot = ifelse(imp_sab_l$incl_var_out == 1, AL_z2_tot, NA))
sab_supl <- df_cohort_imp[[1]] %>% select(all_of(c("HelixID", ap_supl, noise, noise_supl, out_supl)), matches("_z$|_as1$"))
imp_sab_full <- left_join(imp_sab_l, sab_supl, by = "HelixID") %>% as.mids()

imp_ede_l <- complete(imp_ede, "long", include = T)
imp_ede_l <- imp_ede_l %>% 
  mutate(AL_2_tot = ifelse(imp_ede_l$incl_var_out == 1, AL_2_tot, NA),
         AL_z2_tot = ifelse(imp_ede_l$incl_var_out == 1, AL_z2_tot, NA))
ede_supl <- df_cohort_imp[[2]] %>% select(all_of(c("HelixID", "hs_pm25abs_yr_hs_t","hs_pm10_yr_hs_t", ap_supl, noise, noise_supl, out_supl)), matches("_z$|_as1$"))
imp_ede_full <- left_join(imp_ede_l, ede_supl, by = "HelixID") %>%
  relocate("hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t", .after = "hs_pm25_yr_hs_t") %>%
  as.mids()

imp_bib_l <- complete(imp_bib, "long", include = T)
imp_bib_l <- imp_bib_l %>% 
  mutate(AL_2_tot = ifelse(imp_bib_l$incl_var_out == 1, AL_2_tot, NA),
         AL_z2_tot = ifelse(imp_bib_l$incl_var_out == 1, AL_z2_tot, NA))
bib_supl <- df_cohort_imp[[3]] %>% select(all_of(c("HelixID", ap_supl, noise, noise_supl, out_supl)), matches("_z$|_as1$"))
imp_bib_full <- left_join(imp_bib_l, bib_supl, by = "HelixID") %>% as.mids()

imp_rhea_l <- complete(imp_rhea, "long", include = T)
imp_rhea_l <- imp_rhea_l %>% 
  mutate(AL_2_tot = ifelse(imp_rhea_l$incl_var_out == 1, AL_2_tot, NA),
         AL_z2_tot = ifelse(imp_rhea_l$incl_var_out == 1, AL_z2_tot, NA))
rhea_supl <- df_cohort_imp[[4]] %>% select(all_of(c("HelixID", "hs_pm25abs_yr_hs_t", ap_supl, noise, noise_supl, out_supl)), matches("_z$|_as1$"))
imp_rhea_full <- left_join(imp_rhea_l, rhea_supl, by = "HelixID") %>%
  relocate("hs_pm25abs_yr_hs_t", .after = "hs_pm25_yr_hs_t") %>%
  as.mids()

imp_kanc_l <- complete(imp_kanc, "long", include = T)
imp_kanc_l <- imp_kanc_l %>% 
  mutate(AL_2_tot = ifelse(imp_kanc_l$incl_var_out == 1, AL_2_tot, NA),
         AL_z2_tot = ifelse(imp_kanc_l$incl_var_out == 1, AL_z2_tot, NA))
kanc_supl <- df_cohort_imp[[5]] %>% select(all_of(c("HelixID", ap_supl, noise, noise_supl, out_supl)), matches("_z$|_as1$"))
imp_kanc_full <- left_join(imp_kanc_l, kanc_supl, by = "HelixID") %>% as.mids()

imp_moba_l <- complete(imp_moba, "long", include = T)
imp_moba_l <- imp_moba_l %>% 
  mutate(AL_2_tot = ifelse(imp_moba_l$incl_var_out == 1, AL_2_tot, NA),
         AL_z2_tot = ifelse(imp_moba_l$incl_var_out == 1, AL_z2_tot, NA))
moba_supl <- df_cohort_imp[[6]] %>% select(all_of(c("HelixID", ap_supl, noise, noise_supl, out_supl)), matches("_z$|_as1$"))
imp_moba_full <- left_join(imp_moba_l, moba_supl, by = "HelixID") %>% as.mids()

# test run
# test <- mice::rbind(imp_sab_full,imp_ede_full)
#
# test1 <- complete(imp_full, n = 1)
# cova <- test1[,c(cov)]
#
# perc_missings <- function(x){
#   m <- with(test1, sum(is.na(x))/nrow(test1)*100)
# }
# sapply(cova, perc_missings)
# no missing in cov. imputation works.

imp_full <- mice::rbind(imp_sab_full, imp_ede_full) %>%
  rbind(imp_bib_full) %>%
  rbind(imp_rhea_full) %>%
  rbind(imp_kanc_full) %>%
  rbind(imp_moba_full)

save(imp_sab_full, imp_ede_full, imp_bib_full, imp_rhea_full, imp_kanc_full, imp_moba_full, imp_full, file = "/Users/ymou/helix_project/data/analysis_data/imp_full_ds.Rdata")
```
