==> 9_tables.Rmd <==
---
title: "Tables"
author: "Yuchan Mou"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
options(knitr.table.format = "html")
```

```{r load packages, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
# load pkgs
packages <- c("tidyverse", "mice", "readxl", "data.table", "gt", "writexl")
sapply(packages, library, character.only = T)
```

```{r import data}
load("/Users/ymou/helix_project/data/analysis_data/imp_full_ds.Rdata")
# Air pollutants
# Count-based AL, Continuous AL and their weighted average
ap_tot_out <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome2024-03-21.xlsx")
# Count-based AL and weighted average score per system
ap_tot_outps <- read_excel("/Users/ymou/helix_project/results/airpollution_total_outcomepsystem2024-03-21.xlsx")
# Continuous AL and weighted average score per system
ap_tot_outzps <- read_excel("/Users/ymou/helix_project/results/airpollution_total_outcomepsystem_zscore2024-03-21.xlsx")

# Noise
noise_tot_out <- read_excel("/Users/ymou/helix_project/results/noise_mainoutcome2024-03-21.xlsx")

tot_res <- bind_rows(ap_tot_out, ap_tot_outps, ap_tot_outzps, noise_tot_out)
df <- imp_full$data %>% filter(!is.na(AL_2_tot))
analysis_date <- Sys.Date()
```

```{r define variable, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
expo <- c(
  "hs_no2_yr_hs_t", "hs_no2_yr_hs_h", "hs_no2_yr_hs_s",
  "hs_pm25_yr_hs_t", "hs_pm25_yr_hs_h", "hs_pm25_yr_hs_s",
  "hs_pm25abs_yr_hs_t", "hs_pm25abs_yr_hs_h", "hs_pm25abs_yr_hs_s",
  "hs_pm10_yr_hs_t", "hs_pm10_yr_hs_h", "hs_pm10_yr_hs_s",
  "hs_lden_c_h", "hs_lden_c_s"
)

mainexpo <- c( "hs_no2_yr_hs_t", "hs_pm25_yr_hs_t", "hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t", "hs_lden_c_h", "hs_lden_c_s")


# 19 biomarkers included in the main AL score
# hs_non_hdl_c is calculated with hs_totchol_c - hs_hdlchol_c
biomarker_all <- c(
  "hs_zsys_bp", "hs_zdia_bp", "bp_pulse_sd",
  "hs_zbmi_who", "hs_zwaist_mets", "hs_skf_sum2", "hs_fatprop_bia",
  "hs_hdlchol_c", "hs_non_hdl_c", "hs_triglyc_c", "Leptin",
  "Adiponectin",
  "CRP", "IL1beta", "IL6", "IL8", "IL10", "TNFalfa",
  "cortisol_prod_log10"
)

# 4 physiological systems
physystems <- c("cardio", "metab", "immune", "neuro")

cardio <- c("hs_zsys_bp", "hs_zdia_bp", "bp_pulse_sd") # 3 biomarkers
metab <- c(
  "hs_zbmi_who", "hs_zwaist_mets", "hs_skf_sum2", "hs_fatprop_bia",
  "hs_hdlchol_c", "hs_non_hdl_c", "hs_triglyc_c", "Leptin", "Adiponectin"
) # 9 biomarkers
immune <- c("CRP", "IL1beta", "IL6", "IL8", "IL10", "TNFalfa") # 6 biomarkers
neuro <- c("cortisol_prod_log10") # 1 biomarkers

# Outcomes
mainout <- c("AL_2_tot", "AL_z2_tot")

# Covariates
cov <- c(
  "e3_sex", "hs_mvpa", "hs_sd_wk", "hs_child_age_years",
  "FAS_cat", "h_native_2cat", "h_ethnicity_cauc", "h_edumc", "h_edufc",
  "hs_smk_parents", "hs_globalexp_2cat", "e3_asmokyn_p", "e3_alcpreg_yn", "h_marital_2cat", "h_age", "h_parity"
)
catcov <- c(
'cohort', "e3_sex", "FAS_cat", "h_native", "h_ethnicity_cauc", "h_edumc", "h_edufc",
  "hs_smk_parents", "hs_globalexp_2cat", "e3_asmokyn_p", "e3_alcpreg_yn", "h_marital", "h_parity"
)
nonnormalvars <- c("hs_child_age_years", "hs_mvpa", "hs_sd_wk")

```


# Table 1 Allostatic load biomarkers and the cut-off values
```{r tbl1}
df %>%
  select(HelixID, e3_sex, all_of(biomarker_all)) %>%
  group_by(e3_sex) %>% 
  reframe(across(all_of(biomarker_all), ~ quantile(.x, c(0.25, 0.5, 0.75), na.rm = T))) %>% 
  mutate(across(hs_zsys_bp:cortisol_prod_log10, ~round(.,2))) %>% 
  transpose(keep.names = "Biomarkers") %>% 
  rename(
    g_0.25 = V1,
    g_0.5 = V2,
    g_0.75 = V3,
    b_0.25 = V4,
    b_0.5 = V5,
    b_0.75 = V6
  ) %>% 
  slice(-1) %>%  # remove first row
  mutate(across(g_0.25:g_0.75, ~as.numeric(.))) %>% 
  gt() %>% 
  tab_header(
    title = "Table 2 Allostatic load biomarkers and the cut-off values"
  ) %>% 
  fmt_number(decimals = 2, drop_trailing_zeros = F) %>% 
  cols_merge(columns = c(g_0.5, g_0.25, g_0.75),
             pattern = "{1} ({2}, {3})") %>% 
  cols_merge(columns = c(b_0.5, b_0.25, b_0.75),
             pattern = "{1} ({2}, {3})") %>% 
  cols_label(Biomarkers = "Biomarkers",
             g_0.25 = "Girls",
             b_0.25 = "Boys") %>% 
  gt::gtsave(., "/Users/ymou/helix_project/results/Table/Table1-1.docx")

bioz <- paste(biomarker_all, "_z", sep = "")
df %>%
  select(HelixID, e3_sex, matches("_z$")) %>%
  group_by(e3_sex) %>% 
  reframe(across(all_of(bioz), ~ quantile(.x, c(0.25, 0.5, 0.75), na.rm = T))) %>% 
  mutate(across(all_of(bioz), ~round(.,2))) %>% 
  transpose(keep.names = "Biomarkers") %>% 
  rename(
    g_0.25 = V1,
    g_0.5 = V2,
    g_0.75 = V3,
    b_0.25 = V4,
    b_0.5 = V5,
    b_0.75 = V6
  ) %>% 
  slice(-1) %>%  # remove first row
  mutate(across(g_0.25:g_0.75, ~as.numeric(.))) %>% 
  gt() %>% 
  tab_header(
    title = "Table 2 Allostatic load biomarkers and the cut-off values"
  ) %>% 
  fmt_number(decimals = 2, drop_trailing_zeros = F) %>% 
  cols_merge(columns = c(g_0.5, g_0.25, g_0.75),
             pattern = "{1} ({2}, {3})") %>% 
  cols_merge(columns = c(b_0.5, b_0.25, b_0.75),
             pattern = "{1} ({2}, {3})") %>% 
  cols_label(Biomarkers = "Biomarkers",
             g_0.25 = "Girls",
             b_0.25 = "Boys") %>% 
  gt::gtsave(., "/Users/ymou/helix_project/results/Table/Table1-2.docx")

# cortisol 12.5th, 85.5th
df %>%
  select(HelixID, e3_sex, cortisol_prod_log10, cortisol_prod_log10_z) %>%
  group_by(e3_sex) %>% 
  reframe(across(cortisol_prod_log10:cortisol_prod_log10_z, ~ quantile(.x, c(0.125, 0.5, 0.855), na.rm = T))) %>% 
  mutate(across(cortisol_prod_log10:cortisol_prod_log10_z, ~round(.,2))) %>% 
  transpose(keep.names = "Biomarkers") %>% 
  rename(
    g_0.125 = V1,
    g_0.5 = V2,
    g_0.855 = V3,
    b_0.125 = V4,
    b_0.5 = V5,
    b_0.855 = V6
  ) %>% 
  slice(-1) %>%  # remove first row
  mutate(across(g_0.125:g_0.855, ~as.numeric(.))) %>% 
  gt() %>% 
  tab_header(
    title = "Table 2 Allostatic load biomarkers and the cut-off values-cortisol"
  ) %>% 
  fmt_number(decimals = 2, drop_trailing_zeros = F) %>% 
  cols_merge(columns = c(g_0.5, g_0.125, g_0.855),
             pattern = "{1} ({2}, {3})") %>% 
  cols_merge(columns = c(b_0.5, b_0.125, b_0.855),
             pattern = "{1} ({2}, {3})") %>% 
  cols_label(Biomarkers = "Biomarkers",
             g_0.125 = "Girls",
             b_0.125 = "Boys") %>% 
  gt::gtsave(., "/Users/ymou/helix_project/results/Table/Table1-3.docx")
```

# Table 2 Characteristics of Study Population
```{r tbl2}
fct_count(df$cohort, prop = T)

dat_afimp <- complete(imp_full, action = "long", include = F) %>% filter(incl_var_out == 1)

tblone <- CreateTableOne(data = dat_afimp, vars = cov, factorVars = catcov, testNonNormal = kruskal.test)

tbl <- print(tblone, showAllLevels = TRUE, nonnormal = nonnormalvars, contDigits = 1, catDigits = 1, noSpaces = T, format = "fp")
tbl <- as.data.frame(tbl)

write_xlsx(tbl, path = paste("/Users/ymou/helix_project/results/Table/Table2-", "tableone", analysis_date, ".xlsx", sep = ""))

df %>%
  summarise(across(all_of(mainout), ~ mean(.x, na.rm = TRUE)))
df %>%
  reframe(across(all_of(mainout), ~ quantile(.x, c(0.25, 0.5, 0.75), na.rm = T)))

df  %>% 
  summarise(across(all_of(mainout), ~ sd(.x, na.rm = TRUE)))
```

# Table 3 Levels of exposure to outdoor air pollutants and traffic noise in HELIX cohort
```{r tbl3}
median_iqr <- list(
  median = ~median(.x, na.rm = T),
  q25 = ~quantile(.x, 0.25, na.rm = T),
  q75 = ~quantile(.x, 0.75, na.rm = T)
)

# air pollutants
df %>%
  select(cohort, all_of(mainexpo[1:4])) %>%
  group_by(cohort) %>% 
  summarise(across(
    .cols = all_of(mainexpo[1:4]),
    .fns = median_iqr
  )) %>% 
  mutate_at(2:13, round, 1) %>% 
  mutate(
    no2 = paste(hs_no2_yr_hs_t_median, " (", hs_no2_yr_hs_t_q25, ", ", hs_no2_yr_hs_t_q75, ")", sep = ""),
    pm25 = paste(hs_pm25_yr_hs_t_median, " (", hs_pm25_yr_hs_t_q25, ", ", hs_pm25_yr_hs_t_q75, ")", sep = ""),
    pm25abs = paste(hs_pm25abs_yr_hs_t_median, " (", hs_pm25abs_yr_hs_t_q25, ", ", hs_pm25abs_yr_hs_t_q75, ")", sep = ""),
    pm10 = paste(hs_pm10_yr_hs_t_median, " (", hs_pm10_yr_hs_t_q25, ", ", hs_pm10_yr_hs_t_q75, ")", sep = "")
  ) %>% 
  select(cohort, no2, pm25, pm25abs, pm10) %>% 
  transpose(keep.names = "cohort") %>% 
  rename(
    INMA = V1,
    EDEN = V2,
    BIB = V3,
    RHEA = V4,
    KANC = V5,
    MOBA = V6
  ) %>% 
  slice(-1) %>%  # remove first row
  gt() %>% 
  tab_header(
    title = "Table 3 Levels of exposure to outdoor air pollutants in HELIX cohort"
  ) %>% 
  gt::gtsave(., "/Users/ymou/helix_project/results/Table/Table3-1.docx")


# noise 
tblnoise <- CreateTableOne(data = df, vars = mainexpo[5:6], factorVars = mainexpo[5:6], testNonNormal = kruskal.test, strata = "cohort")

tbl <- print(tblnoise, showAllLevels = TRUE, contDigits = 1, catDigits = 1, noSpaces = T, format = "p")
tbl <- as.data.frame(tbl)

write_xlsx(tbl, path = "/Users/ymou/helix_project/results/Table/Table3-2.xlsx")


# allostatic load scores 
tblal <- CreateTableOne(data = df, vars = mainout, testNonNormal = kruskal.test, strata = "cohort")

tbl <- print(tblal, showAllLevels = TRUE, nonnormal = mainout, contDigits = 1, noSpaces = T, format = "fp")
tbl <- as.data.frame(tbl)

write_xlsx(tbl, path = "/Users/ymou/helix_project/results/Table/Table3-3.xlsx")

```


# Table 4 Relative risk of count-based allostatic load score associated with outdoor air pollutants and road traffic noise levels
```{r tbl3}
tot_res %>%
  select(Outcome, Predictor, exp_beta_mod4, exp_ci_mod4) %>%
  filter(Outcome == "AL_2_tot") %>% 
  gt() %>% 
  tab_header(
    title = "Table 3 Relative risk of count-based allostatic load score associated with outdoor air pollutants and road traffic noise levels"
  ) %>% 
  fmt_number(decimals = 2, drop_trailing_zeros = F) %>% 
  gt::gtsave(., "/Users/ymou/helix_project/results/Table/Table4.docx")
```

# Table 5 Effect estimates of continuous allostatic load score associated with outdoor air pollutants and road traffic noise levels
```{r tbl4}
tot_res %>%
  select(Outcome, Predictor, beta_mod4, `95ci_mod4`) %>%
  filter(Outcome == "AL_z2_tot") %>% 
  gt() %>% 
  tab_header(
    title = "Table 4 Effect estimates of continuous allostatic load score associated with outdoor air pollutants and road traffic noise levels"
  ) %>% 
  fmt_number(decimals = 2, drop_trailing_zeros = F) %>% 
  gt::gtsave(., "/Users/ymou/helix_project/results/Table/Table5.docx")
```
