==> 10_resultsplots.Rmd <==
---
title: "Plots"
author: "Yuchan Mou"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
options(knitr.table.format = "html")
```

```{r load packages, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
# load pkgs
packages <- c("tidyverse", "labelled", "openxlsx", "meta", "readxl", "RColorBrewer", "ggsci", "patchwork")
sapply(packages, library, character.only = T)
```

```{r import data, eval=TRUE}
analysis_date <- Sys.Date()

# Air pollutants
# Count-based AL, Continuous AL and their weighted average
ap_tot_out <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome2024-03-21.xlsx")
# Count-based AL and weighted average score per system
ap_tot_outps <- read_excel("/Users/ymou/helix_project/results/airpollution_total_outcomepsystem2024-03-21.xlsx")
# Continuous AL and weighted average score per system
ap_tot_outzps <- read_excel("/Users/ymou/helix_project/results/airpollution_total_outcomepsystem_zscore2024-03-21.xlsx")

# Count-based AL, Continuous AL and their weighted average: Home, school, commuting
ap_hsr_out <- read_excel("/Users/ymou/helix_project/results/airpollution_hsrp_mainoutcome2024-03-21.xlsx") %>% filter(!str_detect(Predictor, "_p$")) # other places not incl
# Count-based AL and weighted average score per system: Home, school, commuting
ap_hsr_outps <- read_excel("/Users/ymou/helix_project/results/airpollution_hsrp_outcomepsystem2024-03-21.xlsx") %>% filter(!str_detect(Predictor, "_p$")) # other places not incl
# Continuous AL and weighted average score per system: Home, school, commuting
ap_hsr_outzps <- read_excel("/Users/ymou/helix_project/results/airpollution_hsrp_outcomepsystem_zscore2024-03-21.xlsx") %>% filter(!str_detect(Predictor, "_p$")) # other places not incl

# Noise
noise_tot_out <- read_excel("/Users/ymou/helix_project/results/noise_mainoutcome2024-03-21.xlsx")
```

```{r merge data}
ap_tot_res <- bind_rows(ap_tot_out, ap_tot_outps, ap_tot_outzps)
ap_tot_res <- ap_tot_res %>%
  mutate(al_type = if_else(str_detect(Outcome, "^AL_2") == T & str_detect(Outcome, "_wavg$") == F, "AL_2",
    if_else(str_detect(Outcome, "^AL_2") == T & str_detect(Outcome, "_wavg$") == T, "AL_2_wavg",
      if_else(str_detect(Outcome, "^AL_z2") == T & str_detect(Outcome, "_wavg$") == F, "AL_z2",
        if_else(str_detect(Outcome, "^AL_z2") == T & str_detect(Outcome, "_wavg$") == T, "AL_z2_wavg", NA)
      )
    )
  )) %>%
  dplyr::rename(
    "beta" = "beta_mod4",
    "low" = "cilo_mod4",
    "hi" = "cihi_mod4"
  ) %>%
  mutate_at(c("Predictor", "Outcome"), as.factor) %>% 
  mutate(Predictor = fct_relevel(Predictor,c("hs_no2_yr_hs_t","hs_pm25_yr_hs_t","hs_pm25abs_yr_hs_t", "hs_pm10_yr_hs_t")),
         Outcome = fct_relevel(Outcome, c("AL_2_tot", "AL_2_cardio", "AL_2_metab", "AL_2_immune", "AL_2_neuro",
                                          "AL_z2_tot", "AL_z2_cardio", "AL_z2_metab", "AL_z2_immune", "AL_z2_neuro",
                                          "AL_2_tot_wavg", "AL_2_cardio_wavg", "AL_2_metab_wavg", "AL_2_immune_wavg", "AL_2_neuro_wavg",
                                          "AL_z2_tot_wavg", "AL_z2_cardio_wavg", "AL_z2_metab_wavg", "AL_z2_immune_wavg", "AL_z2_neuro_wavg"))) %>% 
  mutate(
    Predictor = recode(Predictor,
      "hs_no2_yr_hs_t" = "NO[2]",
      "hs_pm10_yr_hs_t" = "PM[10]",
      "hs_pm25_yr_hs_t" = "PM[2.5]",
      "hs_pm25abs_yr_hs_t" = "PM[2.5][abs]"
    ),
    Outcome = recode(Outcome,
      "AL_2_tot" = "Count-based ALS",
      "AL_2_tot_wavg" = "Count-based weighted average ALS",
      "AL_z2_tot" = "Continuous ALS",
      "AL_z2_tot_wavg" = "Continuous weighted average ALS",
      "AL_2_cardio" = "Count-based cardiovascular ALS",
      "AL_2_metab" = "Count-based metabolic ALS",
      "AL_2_immune" = "Count-based immune/inflammatory ALS",
      "AL_2_neuro" = "Count-based neuroendorine ALS",
      "AL_2_cardio_wavg" = "Count-based cardiovascular weighted average ALS",
      "AL_2_metab_wavg" = "Count-based metabolic weighted average ALS",
      "AL_2_immune_wavg" = "Count-based immune/inflammatory weighted average ALS",
      "AL_2_neuro_wavg" = "Count-based neuroendorine weighted average ALS",
      "AL_z2_cardio" = "Continuous cardiovascular ALS",
      "AL_z2_metab" = "Continuous metabolic ALS",
      "AL_z2_immune" = "Continuous immune/inflammatory ALS",
      "AL_z2_neuro" = "Continuous neuroendorine ALS",
      "AL_z2_cardio_wavg" = "Continuous cardiovascular weighted average ALS",
      "AL_z2_metab_wavg" = "Continuous metabolic weighted average ALS",
      "AL_z2_immune_wavg" = "Continuous immune/inflammatory weighted average ALS",
      "AL_z2_neuro_wavg" = "Continuous neuroendorine weighted average ALS"
    )
  ) %>% 
  rename(Outcomes = Outcome)



ap_hsr_res <- bind_rows(ap_hsr_out, ap_hsr_outps, ap_hsr_outzps)
ap_hsr_res <- ap_hsr_res %>%
  mutate(al_type = if_else(str_detect(Outcome, "^AL_2") == T & str_detect(Outcome, "_wavg$") == F, "AL_2",
    if_else(str_detect(Outcome, "^AL_2") == T & str_detect(Outcome, "_wavg$") == T, "AL_2_wavg",
      if_else(str_detect(Outcome, "^AL_z2") == T & str_detect(Outcome, "_wavg$") == F, "AL_z2",
        if_else(str_detect(Outcome, "^AL_z2") == T & str_detect(Outcome, "_wavg$") == T, "AL_z2_wavg", NA)
      )
    )
  )) %>%
  dplyr::rename(
    "beta" = "beta_mod4",
    "low" = "cilo_mod4",
    "hi" = "cihi_mod4"
  ) %>%
  mutate(type = if_else(str_detect(Predictor, "_h$"), "Home",
    if_else(str_detect(Predictor, "_s$"), "School", "Commuting")
  )) %>%
  mutate_at(c("Predictor", "Outcome"), as.factor) %>% 
  mutate(Predictor = fct_relevel(Predictor,c("hs_no2_yr_hs_h","hs_no2_yr_hs_s","hs_no2_yr_hs_r",
                                             "hs_pm25_yr_hs_h", "hs_pm25_yr_hs_s", "hs_pm25_yr_hs_r",
                                             "hs_pm25abs_yr_hs_h", "hs_pm25abs_yr_hs_s", "hs_pm25abs_yr_hs_r",
                                             "hs_pm10_yr_hs_h", "hs_pm10_yr_hs_s", "hs_pm10_yr_hs_r")),
         Outcome = fct_relevel(Outcome, c("AL_2_tot", "AL_2_cardio", "AL_2_metab", "AL_2_immune", "AL_2_neuro",
                                          "AL_z2_tot", "AL_z2_cardio", "AL_z2_metab", "AL_z2_immune", "AL_z2_neuro",
                                          "AL_2_tot_wavg", "AL_2_cardio_wavg", "AL_2_metab_wavg", "AL_2_immune_wavg", "AL_2_neuro_wavg",
                                          "AL_z2_tot_wavg", "AL_z2_cardio_wavg", "AL_z2_metab_wavg", "AL_z2_immune_wavg", "AL_z2_neuro_wavg"))) %>% 
  mutate(Predictor = recode(Predictor,
    "hs_no2_yr_hs_h" = "NO[2]",
    "hs_no2_yr_hs_s" = "NO[2]",
    "hs_no2_yr_hs_r" = "NO[2]",
    "hs_pm25_yr_hs_h" = "PM[2.5]",
    "hs_pm25_yr_hs_s" = "PM[2.5]",
    "hs_pm25_yr_hs_r" = "PM[2.5]",
    "hs_pm25abs_yr_hs_h" = "PM[2.5][abs]",
    "hs_pm25abs_yr_hs_s" = "PM[2.5][abs]",
    "hs_pm25abs_yr_hs_r" = "PM[2.5][abs]",
    "hs_pm10_yr_hs_h" = "PM[10]",
    "hs_pm10_yr_hs_s" = "PM[10]",
    "hs_pm10_yr_hs_r" = "PM[10]",
  ),
    Outcome = recode(Outcome,
      "AL_2_tot" = "Count-based ALS",
      "AL_2_tot_wavg" = "Count-based weighted average ALS",
      "AL_z2_tot" = "Continuous ALS",
      "AL_z2_tot_wavg" = "Continuous weighted average ALS",
      "AL_2_cardio" = "Count-based cardiovascular ALS",
      "AL_2_metab" = "Count-based metabolic ALS",
      "AL_2_immune" = "Count-based immune/inflammatory ALS",
      "AL_2_neuro" = "Count-based neuroendorine ALS",
      "AL_2_cardio_wavg" = "Count-based cardiovascular weighted average ALS",
      "AL_2_metab_wavg" = "Count-based metabolic weighted average ALS",
      "AL_2_immune_wavg" = "Count-based immune/inflammatory weighted average ALS",
      "AL_2_neuro_wavg" = "Count-based neuroendorine weighted average ALS",
      "AL_z2_cardio" = "Continuous cardiovascular ALS",
      "AL_z2_metab" = "Continuous metabolic ALS",
      "AL_z2_immune" = "Continuous immune/inflammatory ALS",
      "AL_z2_neuro" = "Continuous neuroendorine ALS",
      "AL_z2_cardio_wavg" = "Continuous cardiovascular weighted average ALS",
      "AL_z2_metab_wavg" = "Continuous metabolic weighted average ALS",
      "AL_z2_immune_wavg" = "Continuous immune/inflammatory weighted average ALS",
      "AL_z2_neuro_wavg" = "Continuous neuroendorine weighted average ALS"
    )) %>% 
  rename(Outcomes = Outcome)


noise_tot_out <- noise_tot_out %>%
  mutate(al_type = if_else(str_detect(Outcome, "^AL_2") == T & str_detect(Outcome, "_wavg$") == F, "AL_2",
    if_else(str_detect(Outcome, "^AL_2") == T & str_detect(Outcome, "_wavg$") == T, "AL_2_wavg",
      if_else(str_detect(Outcome, "^AL_z2") == T & str_detect(Outcome, "_wavg$") == F, "AL_z2",
        if_else(str_detect(Outcome, "^AL_z2") == T & str_detect(Outcome, "_wavg$") == T, "AL_z2_wavg", NA)
      )
    )
  )) %>%
  mutate(pred_type = ifelse(str_detect(Predictor, "^hs_lden_c_h") == T, "lden_h", "lden_s")) %>%
  dplyr::rename(
    "beta" = "beta_mod4",
    "low" = "cilo_mod4",
    "hi" = "cihi_mod4"
  ) %>%
  mutate_at(c("Predictor", "pred_type", "Outcome"), as.factor) %>%
  mutate(Predictor = fct_relevel(Predictor,c("hs_lden_c_h_55-59.9","hs_lden_c_h_60-64.9","hs_lden_c_h>=65",
                                             "hs_lden_c_s_55-59.9", "hs_lden_c_s_60-64.9", "hs_lden_c_s>=65")),
         Outcome = fct_relevel(Outcome, c("AL_2_tot", "AL_z2_tot", "AL_2_tot_wavg", "AL_z2_tot_wavg"))) %>% 
  mutate(
    Predictor = recode(Predictor,
      "hs_lden_c_h_55-59.9" = "55-59.9",
      "hs_lden_c_h_60-64.9" = "60-64.9",
      "hs_lden_c_h>=65" = "le65",
      "hs_lden_c_s_55-59.9" = "55-59.9",
      "hs_lden_c_s_60-64.9" = "60-64.9",
      "hs_lden_c_s>=65" = "le65"
    ),
    pred_type = recode(pred_type,
      "lden_h" = "Home",
      "lden_s" = "School"
    ),
    Outcome = recode(Outcome,
      "AL_2_tot" = "Count-based ALS",
      "AL_2_tot_wavg" = "Count-based weighted average ALS",
      "AL_z2_tot" = "Continuous ALS",
      "AL_z2_tot_wavg" = "Continuous weighted average ALS"
    )
  )
```

# Plots

-----------------------------------------------------------------------------------------------------
## Air pollutants

### Count-based AL score
```{r}
dt_p1 <- ap_tot_res %>% filter(al_type == "AL_2")

dt_p1$es <- paste0(
  sprintf("%.2f", dt_p1$exp_beta_mod4), " (",
  sprintf("%.2f", dt_p1$exp_cilo_mod4), "-",
  sprintf("%.2f", dt_p1$exp_cihi_mod4), ")"
)
```

```{r fig.asp = 0.8, fig.width=10}
p1 <- ggplot(
  data = dt_p1,
  aes(x = Outcomes, y = exp_beta_mod4, ymin = exp_cilo_mod4, ymax = exp_cihi_mod4)
) +
  geom_pointrange(aes(col = Outcomes), show.legend = T) +
  geom_hline(yintercept = 1, linetype = 3) +
  xlab("") +
  ylab("Relative risk (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = exp_cilo_mod4, ymax = exp_cihi_mod4, col = Outcomes), width = 0.5, cex = 1) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, scales = "free_y", labeller = labeller(Predictor = label_parsed)) +
  # scale_y_continuous(limits = c(-1, 6), breaks = c(-1, 0, 1, 2, 3, 4, 5, 6)) +
  guides(col = guide_legend(nrow = 2, reverse = T, title = "Count-based ALS",
                            title.position = "top")) +
  
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0, face = "bold"),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 12),
    legend.key=element_blank(),
    legend.position = "bottom",
    legend.justification = "left",
    panel.background = element_rect(fill = 'white', colour = 'black'),
    strip.background=element_rect(fill="white", colour = 'white')
  ) +
  scale_color_npg(
     labels = c("Count-based neuroendorine ALS" = "Neuroendocrine ALS",
               "Count-based immune/inflammatory ALS" = "Immune/inflammatory ALS", 
               "Count-based metabolic ALS" = "Metabolic ALS",
               "Count-based cardiovascular ALS" = "Cardiovascular ALS",
               "Count-based ALS" = "Total ALS")
  ) +
  coord_flip() +
  ggtitle("")

p1

table_es <- ggplot(dt_p1, aes(x = 0, y = Outcomes, label = es)) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, labeller = labeller(Predictor = label_parsed)) +
  geom_text(hjust = 0, size = 5) +
  theme(
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    axis.title.x = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title.y = element_blank(),
    plot.margin = margin(0, 1, 0, 0),
    strip.text.y = element_blank(),
    panel.background = element_rect(fill = 'white', colour = 'white'),
    strip.background=element_rect(fill="white", colour = 'white')
  )
table_es

# Combine plots using patchwork
combined_plot <- p1 + table_es + plot_layout(ncol = 2, widths = c(0.5, 0.5))

# Display the combined plot
print(combined_plot)
```

```{r, eval=F}
ggsave("/Users/ymou/helix_project/results/Fig/fig1-countAL.png", width = 15, height = 10, units = "in", dpi = 300)
```



### Continuous AL score
```{r}
dt_p3 <- ap_tot_res %>% filter(al_type == "AL_z2")

dt_p3$es <- paste0(
  sprintf("%.2f", dt_p3$beta), " (",
  sprintf("%.2f", dt_p3$low), "-",
  sprintf("%.2f", dt_p3$hi), ")"
)
```

```{r fig.asp = 0.8, fig.width=10}
p3 <- ggplot(
  data = dt_p3,
  aes(x = Outcomes, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Outcomes), show.legend = T) +
  geom_hline(yintercept = 0, linetype = 3) +
  xlab("") +
  ylab("Effect sizes (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Outcomes), width = 0.5, cex = 1) +
  guides(color = guide_legend(nrow = 2, reverse = T, title = "Continuous ALS",
                              title.position = "top")) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, scales = "free_y", labeller = labeller(Predictor = label_parsed)) +
  # scale_y_continuous(limits = c(-1, 6), breaks = c(-1, 0, 1, 2, 3, 4, 5, 6)) +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0, face = "bold"),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 12),
    legend.key=element_blank(),
    legend.position = "bottom",
    legend.justification = "left",
    panel.background = element_rect(fill = 'white', colour = 'black'),
    strip.background=element_rect(fill="white", colour = 'white')
  ) +
  scale_color_npg(
    labels = c("Continuous neuroendorine ALS" = "Neuroendocrine ALS",
               "Continuous immune/inflammatory ALS" = "Immune/inflammatory ALS", 
               "Continuous metabolic ALS" = "Metabolic ALS",
               "Continuous cardiovascular ALS" = "Cardiovascular ALS",
               "Continuous ALS" = "Total ALS")
  ) +
  coord_flip() +
  ggtitle("")

p3

table_es <- ggplot(dt_p3, aes(x = 0, y = Outcomes, label = es)) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, labeller = labeller(Predictor = label_parsed)) +
  geom_text(hjust = 0, size = 5) +
  theme(
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    axis.title.x = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title.y = element_blank(),
    plot.margin = margin(0, 1, 0, 0),
    strip.text.y = element_blank(),
    panel.background = element_rect(fill = 'white', colour = 'white'),
    strip.background=element_rect(fill="white", colour = 'white')
  )
table_es

# Combine plots using patchwork
combined_plot <- p3 + table_es + plot_layout(ncol = 2, widths = c(0.5, 0.5))

# Display the combined plot
print(combined_plot)
```

```{r, eval=F}
ggsave("/Users/ymou/helix_project/results/Fig/fig1-continuousAL.png", width = 15, height = 10, units = "in", dpi = 300)
```

-----------------------------------------------------------------------------------------------------
## Air pollutants(home, school, commuting) 

### Count-based AL score
```{r}
dt_p5 <- ap_hsr_res %>% filter(al_type == "AL_2")
```

```{r fig.asp = 0.8, fig.width=10}
p5 <- ggplot(
  data = dt_p5,
  aes(x = Outcomes, y = exp_beta_mod4, ymin = exp_cilo_mod4, ymax = exp_cihi_mod4)
) +
  geom_pointrange(aes(col = Outcomes), show.legend = T) +
  geom_hline(yintercept = 1, linetype = 3) +
  xlab("") +
  ylab("Relative risk (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = exp_cilo_mod4, ymax = exp_cihi_mod4, col = Outcomes), width = 0.5, cex = 1) +
  facet_grid(rows = vars(Predictor), cols = vars(type), switch = "y", labeller = labeller(Predictor = label_parsed)) +
  guides(col = guide_legend(nrow = 2, reverse = T, title = "Count-based ALS",
                            title.position = "top")) +
  # scale_y_continuous(limits = c(-1, 6), breaks = c(-1, 0, 1, 2, 3, 4, 5, 6)) +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text.x = element_text(size = 14, angle = 0),
    strip.text.y.left = element_text(size = 14, angle = 0),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 12),
    legend.key=element_blank(),
    legend.position = "bottom",
    legend.justification = "left",
    panel.background = element_rect(fill = 'white', colour = 'black'),
    strip.background=element_rect(fill="white", colour = 'white')
  ) +
  scale_color_npg(
    labels = c("Count-based neuroendorine ALS" = "Neuroendocrine ALS",
               "Count-based immune/inflammatory ALS" = "Immune/inflammatory ALS", 
               "Count-based metabolic ALS" = "Metabolic ALS",
               "Count-based cardiovascular ALS" = "Cardiovascular ALS",
               "Count-based ALS" = "Total ALS")
  ) +
  coord_flip() +
  ggtitle("")

p5
```


```{r, eval=F}
ggsave("/Users/ymou/helix_project/results/Fig/figs7-countAL-hsr.png", width = 13, height = 13, units = "in", dpi = 300)
```


### Continuous AL score
```{r}
dt_p7 <- ap_hsr_res %>% filter(al_type == "AL_z2")
```

```{r fig.asp = 0.8, fig.width=10}
p7 <- ggplot(
  data = dt_p7,
  aes(x = Outcomes, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Outcomes), show.legend = T) +
  geom_hline(yintercept = 0, linetype = 3) +
  xlab("") +
  ylab("Relative risk (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Outcomes), width = 0.5, cex = 1) +
  facet_grid(rows = vars(Predictor), cols = vars(type), switch = "y", labeller = labeller(Predictor = label_parsed)) +
  guides(col = guide_legend(nrow = 2, reverse = T, title = "Continuous-based ALS",
                            title.position = "top")) +
  # scale_y_continuous(limits = c(-1, 6), breaks = c(-1, 0, 1, 2, 3, 4, 5, 6)) +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text.x = element_text(size = 14, angle = 0),
    strip.text.y.left = element_text(size = 14, angle = 0),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 12),
    legend.key=element_blank(),
    legend.position = "bottom",
    legend.justification = "left",
    panel.background = element_rect(fill = 'white', colour = 'black'),
    strip.background=element_rect(fill="white", colour = 'white')
  ) +
  scale_color_npg(
    labels = c("Continuous neuroendorine ALS" = "Neuroendocrine ALS",
               "Continuous immune/inflammatory ALS" = "Immune/inflammatory ALS", 
               "Continuous metabolic ALS" = "Metabolic ALS",
               "Continuous cardiovascular ALS" = "Cardiovascular ALS",
               "Continuous ALS" = "Total ALS")
  ) +
  coord_flip() +
  ggtitle("")

p7
```

```{r, eval=F}
ggsave("/Users/ymou/helix_project/results/Fig/figs7-continousAL-hsr.png", width = 13, height = 13, units = "in", dpi = 300)
```


-----------------------------------------------------------------------------------------------------
## Road traffic noise

### Count-based AL score
```{r}
dt_p9 <- noise_tot_out %>% filter(al_type == "AL_2")

dt_p9$es <- paste0(
  sprintf("%.2f", dt_p9$exp_beta_mod4), " (",
  sprintf("%.2f", dt_p9$exp_cilo_mod4), "-",
  sprintf("%.2f", dt_p9$exp_cihi_mod4), ")"
)
dt_p9$pred_tbl <- paste0(dt_p9$pred_type, dt_p9$Predictor)
```

```{r fig.asp = 0.8, fig.width=10}
p9 <- ggplot(
  data = dt_p9,
  aes(x = Predictor, y = exp_beta_mod4, ymin = exp_cilo_mod4, ymax = exp_cihi_mod4)
) +
  geom_pointrange(aes(col = Predictor), show.legend = T) +
  geom_hline(yintercept = 1, linetype = 3) +
  scale_x_discrete(limits = rev(levels(dt_p9$Predictor))) +
  xlab("") +
  ylab("Relative risk (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = exp_cilo_mod4, ymax = exp_cihi_mod4, col = Predictor), width = 0.5, cex = 1) +
  facet_wrap(~pred_type, strip.position = "left", nrow = 9, scales = "free_y") +
  # scale_y_continuous(limits = c(-1, 6), breaks = c(-1, 0, 1, 2, 3, 4, 5, 6)) +
  guides(col = guide_legend(nrow = 1, reverse = F, title = "Road traffic Noise levels (dB)",
                            title.position = "top")) +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 12),
    legend.key=element_blank(),
    legend.position = "bottom",
    legend.justification = "left",
    panel.background = element_rect(fill = 'white', colour = 'black'),
    strip.background=element_rect(fill="white", colour = 'white')
  ) +
  scale_color_npg(
    labels = c("55to59.9" = "55-<60",
               "60to64.9" = "60-<65",
               "le65" = "≥65")
  ) +
  coord_flip() +
  ggtitle("")

p9

table_es <- ggplot(dt_p9, aes(x = 0, y = Outcome, label = es)) +
  facet_wrap(~pred_tbl, strip.position = "left", nrow = 6) +
  geom_text(hjust = 0, size = 5) +
  theme(
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    axis.title.x = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title.y = element_blank(),
    plot.margin = margin(0, 1, 0, 0),
    strip.text.y = element_blank(),
    panel.background = element_rect(fill = 'white', colour = 'white'),
    strip.background=element_rect(fill="white", colour = 'white')
  )
table_es

# Combine plots using patchwork
combined_plot <- p9 + table_es + plot_layout(ncol = 2, widths = c(0.5, 0.5))

# Display the combined plot
print(combined_plot)
```

```{r, eval=F}
ggsave("/Users/ymou/helix_project/results/Fig/fig1-countAL-noise.png", width = 15, height = 4.2, units = "in", dpi = 300)
```

### Continuous AL score
```{r}
dt_p10 <- noise_tot_out %>% filter(al_type == "AL_z2")

dt_p10$es <- paste0(
  sprintf("%.2f", dt_p10$beta), " (",
  sprintf("%.2f", dt_p10$low), "-",
  sprintf("%.2f", dt_p10$hi), ")"
)
dt_p10$pred_tbl <- paste0(dt_p10$pred_type, dt_p10$Predictor)
```

```{r fig.asp = 0.8, fig.width=10}
p10 <- ggplot(
  data = dt_p10,
  aes(x = Predictor, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Predictor), show.legend = T) +
  geom_hline(yintercept = 0, linetype = 3) +
  scale_x_discrete(limits = rev(levels(dt_p10$Predictor))) +
  xlab("") +
  ylab("Effect sizes (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Predictor), width = 0.5, cex = 1) +
  guides(color = guide_legend(reverse = TRUE)) +
  facet_wrap(~pred_type, strip.position = "left", nrow = 9, scales = "free_y") +
  # scale_y_continuous(limits = c(-1, 6), breaks = c(-1, 0, 1, 2, 3, 4, 5, 6)) +
  guides(col = guide_legend(nrow = 1, reverse = F, title = "Road traffic Noise levels (dB)",
                            title.position = "top")) +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 12),
    legend.key=element_blank(),
    legend.position = "bottom",
    legend.justification = "left",
    panel.background = element_rect(fill = 'white', colour = 'black'),
    strip.background=element_rect(fill="white", colour = 'white')
  ) +
  scale_color_npg(
    labels = c("55to59.9" = "55-<60",
               "60to64.9" = "60-<65",
               "le65" = "≥65")
  ) +
  coord_flip() +
  ggtitle("")

p10

table_es <- ggplot(dt_p10, aes(x = 0, y = Outcome, label = es)) +
  facet_wrap(~pred_tbl, strip.position = "left", nrow = 6) +
  geom_text(hjust = 0, size = 5) +
  theme(
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    axis.title.x = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title.y = element_blank(),
    plot.margin = margin(0, 1, 0, 0),
    strip.text.y = element_blank(),
    panel.background = element_rect(fill = 'white', colour = 'white'),
    strip.background=element_rect(fill="white", colour = 'white')
  )
table_es

# Combine plots using patchwork
combined_plot <- p10 + table_es + plot_layout(ncol = 2, widths = c(0.5, 0.5))

# Display the combined plot
print(combined_plot)
```

```{r, eval=F}
ggsave("/Users/ymou/helix_project/results/Fig/fig1-continuousAL-noise.png", width = 15, height = 4.2, units = "in", dpi = 300)
```
