==> 11_meta-analysis_forest_plot.Rmd <==
---
title: "Meta-analysis"
author: "Yuchan Mou"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
options(knitr.table.format = "html")
```

```{r load packages, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
# load pkgs
packages <- c("tidyverse", "openxlsx", "meta", "readxl", "RColorBrewer")
sapply(packages, library, character.only = T)
```

```{r, eval=T}
ap <- c(
  "hs_no2_yr_hs_t",
  "hs_pm25_yr_hs_t",
  "hs_pm25abs_yr_hs_t",
  "hs_pm10_yr_hs_t"
)

out <- c("AL_2_tot", "AL_2_tot_wavg", "AL_z2_tot", "AL_z2_tot_wavg")
```

```{r import data, eval=T}
analysis_date <- Sys.Date()
sab <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome_sab2024-03-21.xlsx")
ede <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome_ede2024-03-21.xlsx")
bib <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome_bib2024-03-21.xlsx")
rhea <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome_rhea2024-03-21.xlsx")
kanc <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome_kanc2024-03-21.xlsx")
moba <- read_excel("/Users/ymou/helix_project/results/airpollution_total_mainoutcome_moba2024-03-21.xlsx")
meta_n <- read_excel("/Users/ymou/helix_project/results/meta_n2024-03-21.xlsx")
```

```{r select var for meta, eval=T}
# Only selected model 4
sab <- sab %>%
  mutate(cohort = "sab") %>%
  select(cohort, Outcome, Predictor, beta_mod4, std_mod4, cilo_mod4, cihi_mod4)
ede <- ede %>%
  mutate(cohort = "eden") %>%
  select(cohort, Outcome, Predictor, beta_mod4, std_mod4, cilo_mod4, cihi_mod4)
bib <- bib %>%
  mutate(cohort = "bib") %>%
  select(cohort, Outcome, Predictor, beta_mod4, std_mod4, cilo_mod4, cihi_mod4)
rhea <- rhea %>%
  mutate(cohort = "rhea") %>%
  select(cohort, Outcome, Predictor, beta_mod4, std_mod4, cilo_mod4, cihi_mod4)
kanc <- kanc %>%
  mutate(cohort = "kanc") %>%
  select(cohort, Outcome, Predictor, beta_mod4, std_mod4, cilo_mod4, cihi_mod4)
moba <- moba %>%
  mutate(cohort = "moba") %>%
  select(cohort, Outcome, Predictor, beta_mod4, std_mod4, cilo_mod4, cihi_mod4)

df <- rbind(sab, ede, bib, rhea, kanc, moba)
```

```{r calc combined effect and se, eval=T}
meta_tbl <- as.data.frame(matrix(NA, nrow = 1, ncol = 4))
colnames(meta_tbl) <- c(
  "Outcome", "Predictor",
  "c_std.error", "c_beta"
)

count <- 0
for (o in out) {
  for (d in ap) {
    sep_coefs <- df %>%
      filter(Outcome == o) %>%
      filter(Predictor == d)

    sep_coefs <- sep_coefs %>%
      mutate(
        w = 1 / (std_mod4)^2
      )

    count <- count + 1
    if (d %in% c("hs_no2_yr_hs_t", "hs_pm25_yr_hs_t")) {
      meta_tbl[count, 1] <- o
      meta_tbl[count, 2] <- d
      meta_tbl[count, 3] <- sqrt(1 / (sum(sep_coefs$w)))
      meta_tbl[count, 4] <- (sep_coefs$w[1] * sep_coefs$beta_mod4[1] +
        sep_coefs$w[2] * sep_coefs$beta_mod4[2] +
        sep_coefs$w[3] * sep_coefs$beta_mod4[3] +
        sep_coefs$w[4] * sep_coefs$beta_mod4[4] +
        sep_coefs$w[5] * sep_coefs$beta_mod4[5] +
        sep_coefs$w[6] * sep_coefs$beta_mod4[6]) / (sum(sep_coefs$w))
    }
    if (d %in% c("hs_pm10_yr_hs_t")) {
      meta_tbl[count, 1] <- o
      meta_tbl[count, 2] <- d
      meta_tbl[count, 3] <- sqrt(1 / (sum(sep_coefs$w)))
      meta_tbl[count, 4] <- (sep_coefs$w[1] * sep_coefs$beta_mod4[1] +
        sep_coefs$w[2] * sep_coefs$beta_mod4[2] +
        sep_coefs$w[3] * sep_coefs$beta_mod4[3] +
        sep_coefs$w[4] * sep_coefs$beta_mod4[4] +
        sep_coefs$w[5] * sep_coefs$beta_mod4[5]) / (sum(sep_coefs$w))
    }
    if (d %in% c("hs_pm25abs_yr_hs_t")) {
      meta_tbl[count, 1] <- o
      meta_tbl[count, 2] <- d
      meta_tbl[count, 3] <- sqrt(1 / (sum(sep_coefs$w)))
      meta_tbl[count, 4] <- (sep_coefs$w[1] * sep_coefs$beta_mod4[1] +
        sep_coefs$w[2] * sep_coefs$beta_mod4[2] +
        sep_coefs$w[3] * sep_coefs$beta_mod4[3] +
        sep_coefs$w[4] * sep_coefs$beta_mod4[4]) / (sum(sep_coefs$w))
    }
  }
}

meta_tbl <- meta_tbl %>%
  mutate(
    cohort = "total"
  ) %>%
  relocate(cohort, .before = "Outcome") %>%
  relocate(c_beta, .before = "c_std.error")
```

```{r merge to meta table, eval=TRUE}
df <- df %>%
  dplyr::rename(
    beta = beta_mod4,
    se = std_mod4,
    low = cilo_mod4,
    hi = cihi_mod4
  ) %>%
  mutate(
    weight = 1 / (se)^2
  )

meta_tbl <- meta_tbl %>%
  dplyr::rename(
    beta = c_beta,
    se = c_std.error
  )

df_meta <- bind_rows(df, meta_tbl) # bind_row combine datasets with different numbers of columns
df_meta <- df_meta %>%
  arrange(Outcome, Predictor)

df_meta <- df_meta %>%
  left_join(meta_n, by = c("cohort", "Outcome", "Predictor"))
```

```{r, eval=FALSE}
write.xlsx(df_meta, file = paste("/Users/ymou/helix_project/results/", "meta_tbl", analysis_date, ".xlsx", sep = ""), rowNames = FALSE, colNames = TRUE)
```

```{r data manipulation, eval = F}
# fixed effect model
df_meta <- df_meta %>%
  # mutate(hi = if_else(cohort == "total", beta + 1.96 * se / sqrt(n), df_meta$hi, NA),
  #        low = if_else(cohort == "total", beta - 1.96 * se / sqrt(n), df_meta$low, NA)) %>%
  mutate_at("Predictor", as_factor) %>%
  mutate(Predictor = recode(Predictor,
    "hs_no2_yr_hs_t" = "NO[2]",
    "hs_pm25_yr_hs_t" = "PM[2.5]",
    "hs_pm25abs_yr_hs_t" = "PM[2.5][abs]",
    "hs_pm10_yr_hs_t" = "PM[10]"
  )) %>%
  mutate(Cohort = fct_relevel(
    cohort,
    "total", "sab", "eden", "bib",
    "rhea", "kanc", "moba"
  )) %>%
  mutate(Cohort = recode(Cohort,
    "total" = "Summary effect size",
    "sab" = "INMA",
    "eden" = "EDEN",
    "bib" = "BIB",
    "rhea" = "RHEA",
    "kanc" = "KANC",
    "moba" = "MOBA"
  ))
```

## Forest plot
```{r forest plot, eval = F, fig.asp = 0.8, fig.width=9}
## Fixed effect model
library("ggsci")

df_meta_AL2 <- df_meta %>%
  filter(Outcome %in% c("AL_2_tot"))

p <- ggplot(
  data = df_meta_AL2,
  aes(x = Cohort, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Cohort)) +
  geom_hline(yintercept = 0, linetype = 2) +
  xlab("") +
  ylab("Effect sizes (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Cohort), width = 0.5, cex = 1) +
  scale_color_npg() +
  guides(color = guide_legend(reverse = TRUE)) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, scales = "free_y", labeller = labeller(Predictor = label_parsed)) +
  theme(
    plot.title = element_text(size = 14, face = "bold"),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(face = "bold"),
    axis.title = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(angle = 0, face = "bold")
  ) +
  coord_flip() +
  ggtitle("Cut-off based allostatic load score")


p

df_meta_AL2 <- df_meta %>%
  filter(Outcome %in% c("AL_2_tot_wavg"))

p <- ggplot(
  data = df_meta_AL2,
  aes(x = Cohort, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Cohort)) +
  xlab("") +
  geom_hline(yintercept = 0, linetype = 2) +
  ylab("Effect sizes (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Cohort), width = 0.5, cex = 1) +
  scale_color_npg() +
  guides(color = guide_legend(reverse = TRUE)) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, scales = "free_y", labeller = labeller(Predictor = label_parsed)) +
  theme(
    plot.title = element_text(size = 14, face = "bold"),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(face = "bold"),
    axis.title = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(angle = 0, face = "bold")
  ) +
  coord_flip() +
  ggtitle("Cut-off based weighted average allostatic load score")

p


df_meta_AL2 <- df_meta %>%
  filter(Outcome %in% c("AL_z2_tot"))

p <- ggplot(
  data = df_meta_AL2,
  aes(x = Cohort, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Cohort)) +
  geom_hline(yintercept = 0, linetype = 2) +
  xlab("") +
  ylab("Effect sizes (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Cohort), width = 0.5, cex = 1) +
  scale_color_npg() +
  guides(color = guide_legend(reverse = TRUE)) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, scales = "free_y", labeller = labeller(Predictor = label_parsed)) +
  theme(
    plot.title = element_text(size = 14, face = "bold"),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(face = "bold"),
    axis.title = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(angle = 0, face = "bold")
  ) +
  coord_flip() +
  ggtitle("Z-score based allostatic load score")

p

df_meta_AL2 <- df_meta %>%
  filter(Outcome %in% c("AL_z2_tot_wavg"))

p <- ggplot(
  data = df_meta_AL2,
  aes(x = Cohort, y = beta, ymin = low, ymax = hi)
) +
  geom_pointrange(aes(col = Cohort)) +
  geom_hline(yintercept = 0, linetype = 2) +
  xlab("") +
  ylab("Effect sizes (95% Confidence Interval)") +
  geom_errorbar(aes(ymin = low, ymax = hi, col = Cohort), width = 0.5, cex = 1) +
  scale_color_npg() +
  guides(color = guide_legend(reverse = TRUE)) +
  facet_wrap(~Predictor, strip.position = "left", nrow = 9, scales = "free_y", labeller = labeller(Predictor = label_parsed)) +
  theme(
    plot.title = element_text(size = 14, face = "bold"),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_text(face = "bold"),
    axis.title = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(angle = 0, face = "bold")
  ) +
  coord_flip() +
  ggtitle("Z-score based weighted average allostatic load score")

p
```

## Random effect models

### Count-based AL
```{r, eval=T, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot" & Predictor == "NO[2]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

# beta to RR
# each cohort
test$TE <- exp(test$TE)
test$lower <- exp(test$lower)
test$upper <- exp(test$upper)
# overall te
test$TE.random <- exp(test$TE.random)
test$lower.random <- exp(test$lower.random)
test$upper.random <- exp(test$upper.random)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-no2.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Relative risk", "[95% CI]"),
  xlim = c(0, 2),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "NO2 and count-based allostatic load score",
  fs.smlab = 12
)
dev.off()

```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot" & Predictor == "PM[10]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

# beta to RR
# each cohort
test$TE <- exp(test$TE)
test$lower <- exp(test$lower)
test$upper <- exp(test$upper)
# overall te
test$TE.random <- exp(test$TE.random)
test$lower.random <- exp(test$lower.random)
test$upper.random <- exp(test$upper.random)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-pm10.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Relative risk", "[95% CI]"),
  xlim = c(0, 2),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM10 and count-based allostatic load score",
  fs.smlab = 12
)
dev.off()
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot" & Predictor == "PM[2.5]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

# beta to RR
# each cohort
test$TE <- exp(test$TE)
test$lower <- exp(test$lower)
test$upper <- exp(test$upper)
# overall te
test$TE.random <- exp(test$TE.random)
test$lower.random <- exp(test$lower.random)
test$upper.random <- exp(test$upper.random)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-pm25.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Relative risk", "[95% CI]"),
  xlim = c(0, 2),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5 and count-based allostatic load score",
  fs.smlab = 12
)
dev.off()
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot" & Predictor == "PM[2.5][abs]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

# beta to RR
# each cohort
test$TE <- exp(test$TE)
test$lower <- exp(test$lower)
test$upper <- exp(test$upper)
# overall te
test$TE.random <- exp(test$TE.random)
test$lower.random <- exp(test$lower.random)
test$upper.random <- exp(test$upper.random)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-pm25abs.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Relative risk", "[95% CI]"),
  xlim = c(0, 2),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5abs and count-based allostatic load score",
  fs.smlab = 12
)
dev.off()
```

### Count-based weighted average AL

```{r, eval=T, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot_wavg" & Predictor == "NO[2]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "NO2 and count-based wavg allostatic load score",
  fs.smlab = 12
)
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot_wavg" & Predictor == "PM[10]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM10 and count-based wavg allostatic load score",
  fs.smlab = 12
)
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot_wavg" & Predictor == "PM[2.5]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5 and count-based wavg allostatic load score",
  fs.smlab = 12
)
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_2_tot_wavg" & Predictor == "PM[2.5][abs]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5abs and count-based wavg allostatic load score",
  fs.smlab = 12
)
```

### Continuous AL
```{r, eval=T, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot" & Predictor == "NO[2]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-no2_alz.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "NO2 and continuous allostatic load score",
  fs.smlab = 12
)
dev.off()
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot" & Predictor == "PM[10]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-pm10-alz.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM10 and continuous allostatic load score",
  fs.smlab = 12
)
dev.off()
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot" & Predictor == "PM[2.5]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-pm25-alz.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5 and continuous allostatic load score",
  fs.smlab = 12
)
dev.off()
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot" & Predictor == "PM[2.5][abs]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

png(
  file = "/Users/ymou/helix_project/results/Fig/meta-pm25abs-alz.png",
  width = 10,
  height = 10,
  units = "in",
  res = 1200
)
forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-2.5, 2.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5abs and continuous allostatic load score",
  fs.smlab = 12
)
dev.off()
```

### Continuous weighted average AL
```{r, eval=T, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot_wavg" & Predictor == "NO[2]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-0.5, 0.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "NO2 and continuous wavg allostatic load score",
  fs.smlab = 12
)
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot_wavg" & Predictor == "PM[10]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-0.5, 0.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM10 and continuous wavg allostatic load score",
  fs.smlab = 12
)
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot_wavg" & Predictor == "PM[2.5]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-0.5, 0.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5 and continuous wavg allostatic load score",
  fs.smlab = 12
)
```

```{r, fig.asp = 0.8, fig.width=9}
df_test <- df_meta %>%
  filter(Outcome == "AL_z2_tot_wavg" & Predictor == "PM[2.5][abs]") %>%
  filter(Cohort != "Summary effect size")
test <- metagen(
  TE = beta,
  seTE = se,
  data = df_test,
  sm = "Estimate",
  studlab = Cohort,
  n.e = n,
  random = TRUE
)

forest(test,
  rightcols = c("effect", "ci"),
  leftcols = c("studlab"),
  leftlabs = "Cohorts",
  rightlabs = c("Coefficient", "[95% CI]"),
  xlim = c(-0.5, 0.5),
  comb.fixed = F,
  fontsize = 12,
  digits = 2,
  print.I2 = T,
  print.tau2 = F,
  print.pval = T,
  common = F,
  col.random = "red",
  col.diamond = "olivedrab3",
  plotwidth = "10cm",
  colgap.forest.left = "2cm",
  smlab = "PM2.5abs and continuous wavg allostatic load score",
  fs.smlab = 12
)
```
